refactor(scanner): split scanner_impl.rs into per-token-family modules
scanner_impl.rs was over the 2000-line hard limit (§19). Extract the
token-family scan helpers and the unit tests out of the monolithic wasm
impl block into sibling child modules under scanner_impl/: strings,
numbers, identifiers, templates, slash, jsx, jsdoc, and tests.

Behavior-preserving: methods stay inherent on ScannerState via per-module
impl blocks. Families that export JS bindings keep their #[wasm_bindgen]
impl so every pub fn stays exported; families with no exported methods use
a plain impl. Only helpers genuinely called across module boundaries are
widened to pub(crate); methods used only within their own module stay
private. No new clippy suppressions are introduced, so the workspace
suppression cap (#9446) is preserved.

scanner_impl.rs drops to ~1463 lines and every shard is well under the
limit.

Closes #9431
