+++ crates/tsz-scanner/src/scanner_impl.rs
@@ -13,6 +13,14 @@ use tsz_common::diagnostics::{diagnostic_codes, diagnostic_messages};
 use tsz_common::interner::{Atom, Interner};
 use wasm_bindgen::prelude::wasm_bindgen;
 
+mod identifiers;
+mod jsdoc;
+mod jsx;
+mod numbers;
+mod slash;
+mod strings;
+mod templates;
+
 // =============================================================================
 // Token Flags
 // =============================================================================
@@ -1022,2003 +1030,6 @@ impl ScannerState {
             }
         }
     }
-
-    /// Scan a string literal.
-    fn scan_string(&mut self, quote: u32) {
-        self.pos += 1; // Skip opening quote
-        let mut result = String::new();
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == quote {
-                self.pos += 1; // Closing quote
-                self.token_value = result;
-                self.token = SyntaxKind::StringLiteral;
-                return;
-            }
-            if ch == CharacterCodes::BACKSLASH {
-                self.pos += 1;
-                if self.pos < self.end {
-                    self.scan_string_escape(quote, &mut result);
-                } else {
-                    // Backslash at EOF — incomplete escape sequence.
-                    // Mark with UnterminatedAtEof so the parser can emit TS1126
-                    // instead of TS1002.
-                    self.token_flags |=
-                        TokenFlags::Unterminated as u32 | TokenFlags::UnterminatedAtEof as u32;
-                    self.token_value = result;
-                    self.token = SyntaxKind::StringLiteral;
-                    return;
-                }
-            } else if ch == CharacterCodes::LINE_FEED || ch == CharacterCodes::CARRIAGE_RETURN {
-                // Unterminated string
-                self.token_flags |= TokenFlags::Unterminated as u32;
-                self.token_value = result;
-                self.token = SyntaxKind::StringLiteral;
-                return;
-            } else {
-                if let Some(c) = char::from_u32(ch) {
-                    result.push(c);
-                }
-                self.pos += self.char_len_at(self.pos); // Advance by character byte length
-            }
-        }
-
-        // Unterminated string — reached EOF without closing quote.
-        // If the string contains an invalid/incomplete escape sequence (e.g., \u{...
-        // without closing brace, or bare \u at EOF), tsc emits TS1126 "Unexpected end
-        // of text". Otherwise, tsc emits TS1002 "Unterminated string literal".
-        self.token_flags |= TokenFlags::Unterminated as u32;
-        if (self.token_flags & TokenFlags::ContainsInvalidEscape as u32) != 0 {
-            self.token_flags |= TokenFlags::UnterminatedAtEof as u32;
-        }
-        self.token_value = result;
-        self.token = SyntaxKind::StringLiteral;
-    }
-
-    fn scan_string_escape(&mut self, quote: u32, result: &mut String) {
-        // pos currently points to the char after the backslash; backslash is at pos - 1.
-        let backslash_pos = self.pos - 1;
-        let escaped = self.char_code_unchecked(self.pos);
-        // Get byte length of the escaped character before advancing.
-        let escaped_len = self.char_len_at(self.pos);
-        self.pos += escaped_len;
-
-        match escaped {
-            CharacterCodes::_0 => self.scan_string_escape_zero(backslash_pos, result),
-            CharacterCodes::_1
-            | CharacterCodes::_2
-            | CharacterCodes::_3
-            | CharacterCodes::_4
-            | CharacterCodes::_5
-            | CharacterCodes::_6
-            | CharacterCodes::_7 => self.scan_string_escape_octal(backslash_pos, escaped, result),
-            CharacterCodes::_8 | CharacterCodes::_9 => {
-                // \8 or \9 is not a valid escape sequence — emit TS1488.
-                let digit = char::from_u32(escaped).unwrap_or('?');
-                self.scanner_diagnostics.push(ScannerDiagnostic {
-                    pos: backslash_pos,
-                    length: self.pos - backslash_pos,
-                    message: diagnostic_messages::ESCAPE_SEQUENCE_IS_NOT_ALLOWED,
-                    code: diagnostic_codes::ESCAPE_SEQUENCE_IS_NOT_ALLOWED,
-                    args: vec![format!("\\{digit}")],
-                });
-                if let Some(c) = char::from_u32(escaped) {
-                    result.push(c);
-                }
-            }
-            CharacterCodes::LOWER_N => result.push('\n'),
-            CharacterCodes::LOWER_R => result.push('\r'),
-            CharacterCodes::LOWER_T => result.push('\t'),
-            CharacterCodes::LOWER_V => result.push('\x0B'),
-            CharacterCodes::LOWER_B => result.push('\x08'),
-            CharacterCodes::LOWER_F => result.push('\x0C'),
-            CharacterCodes::BACKSLASH => result.push('\\'),
-            c if c == quote => result.push(char::from_u32(quote).unwrap_or('\0')),
-            CharacterCodes::LOWER_X => self.scan_string_escape_hex(result),
-            CharacterCodes::LOWER_U => self.scan_string_escape_unicode(result),
-            CharacterCodes::LINE_FEED
-            | CharacterCodes::CARRIAGE_RETURN
-            | CharacterCodes::LINE_SEPARATOR
-            | CharacterCodes::PARAGRAPH_SEPARATOR => {
-                // Line continuation - also handle CR+LF as a single line break
-                if escaped == CharacterCodes::CARRIAGE_RETURN
-                    && self.pos < self.end
-                    && self.char_code_unchecked(self.pos) == CharacterCodes::LINE_FEED
-                {
-                    self.pos += 1;
-                }
-            }
-            _ => {
-                if let Some(c) = char::from_u32(escaped) {
-                    result.push(c);
-                }
-            }
-        }
-    }
-
-    fn scan_string_escape_zero(&mut self, backslash_pos: usize, result: &mut String) {
-        if self.pos < self.end && is_digit(self.char_code_unchecked(self.pos)) {
-            // Legacy octal escape: \0N - scan as octal
-            let mut value = 0u32;
-            let octal_start = self.pos - 1; // include the '0'
-            self.pos = octal_start;
-            while self.pos < self.end
-                && self.pos < octal_start + 3
-                && is_octal_digit(self.char_code_unchecked(self.pos))
-            {
-                value = value * 8 + (self.char_code_unchecked(self.pos) - CharacterCodes::_0);
-                self.pos += 1;
-            }
-            self.emit_octal_escape_diagnostic(backslash_pos, value);
-            if let Some(c) = char::from_u32(value) {
-                result.push(c);
-            }
-        } else {
-            result.push('\0');
-        }
-    }
-
-    fn scan_string_escape_octal(
-        &mut self,
-        backslash_pos: usize,
-        escaped: u32,
-        result: &mut String,
-    ) {
-        // Legacy octal escape: \1 through \7.
-        // tsc reads up to 3 digits when the leading digit is 0-3, but only up to
-        // 2 digits when the leading digit is 4-7 (so `\477` parses as `\47` + '7').
-        let mut value = escaped - CharacterCodes::_0;
-        let max_digits = if escaped <= CharacterCodes::_3 { 3 } else { 2 };
-        let mut count = 1;
-        while count < max_digits
-            && self.pos < self.end
-            && is_octal_digit(self.char_code_unchecked(self.pos))
-        {
-            value = value * 8 + (self.char_code_unchecked(self.pos) - CharacterCodes::_0);
-            self.pos += 1;
-            count += 1;
-        }
-        self.emit_octal_escape_diagnostic(backslash_pos, value);
-        if let Some(c) = char::from_u32(value) {
-            result.push(c);
-        }
-    }
-
-    fn emit_octal_escape_diagnostic(&mut self, backslash_pos: usize, value: u32) {
-        // tsc renders the suggestion as `\xNN` with lowercase 2-digit hex.
-        let suggestion = format!("\\x{value:02x}");
-        self.scanner_diagnostics.push(ScannerDiagnostic {
-            pos: backslash_pos,
-            length: self.pos - backslash_pos,
-            message: diagnostic_messages::OCTAL_ESCAPE_SEQUENCES_ARE_NOT_ALLOWED_USE_THE_SYNTAX,
-            code: diagnostic_codes::OCTAL_ESCAPE_SEQUENCES_ARE_NOT_ALLOWED_USE_THE_SYNTAX,
-            args: vec![suggestion],
-        });
-    }
-
-    fn scan_string_escape_hex(&mut self, result: &mut String) {
-        // Mirror tsc's `scanHexDigits(/*minCount*/ 2)` loop: consume hex digits
-        // one at a time, stop at the first non-hex char, then if fewer than 2
-        // were consumed, emit "Hexadecimal digit expected." at the current
-        // position. tsc's error anchor is wherever the scan halted — the first
-        // non-hex char or the closing quote — not the start of the escape.
-        let mut digit_count = 0;
-        while digit_count < 2
-            && self.pos < self.end
-            && is_hex_digit(self.char_code_unchecked(self.pos))
-        {
-            self.pos += 1;
-            digit_count += 1;
-        }
-        if digit_count < 2 {
-            self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-            self.scanner_diagnostics.push(ScannerDiagnostic {
-                pos: self.pos,
-                length: 0,
-                args: Vec::new(),
-                message: diagnostic_messages::HEXADECIMAL_DIGIT_EXPECTED,
-                code: diagnostic_codes::HEXADECIMAL_DIGIT_EXPECTED,
-            });
-            result.push('\\');
-            result.push('x');
-            return;
-        }
-        let hex_start = self.pos - 2;
-        let hex = self.substring(hex_start, self.pos);
-        if let Ok(code) = u32::from_str_radix(&hex, 16)
-            && let Some(c) = char::from_u32(code)
-        {
-            result.push(c);
-        }
-    }
-
-    fn scan_string_escape_unicode(&mut self, result: &mut String) {
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::OPEN_BRACE {
-            self.pos += 1;
-            let hex_start = self.pos;
-            while self.pos < self.end && is_hex_digit(self.char_code_unchecked(self.pos)) {
-                self.pos += 1;
-            }
-            if self.pos < self.end
-                && self.char_code_unchecked(self.pos) == CharacterCodes::CLOSE_BRACE
-            {
-                let hex = self.substring(hex_start, self.pos);
-                self.pos += 1;
-                if let Ok(code) = u32::from_str_radix(&hex, 16)
-                    && let Some(c) = char::from_u32(code)
-                {
-                    result.push(c);
-                    return;
-                }
-            }
-            // Invalid or unterminated \u{...} escape
-            self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-            result.push('\\');
-            result.push('u');
-            return;
-        }
-        if self.pos + 4 <= self.end {
-            let hex = self.substring(self.pos, self.pos + 4);
-            if let Ok(code) = u32::from_str_radix(&hex, 16)
-                && let Some(c) = char::from_u32(code)
-            {
-                self.pos += 4;
-                result.push(c);
-                return;
-            }
-            // Invalid \uXXXX escape
-            self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-            result.push('\\');
-            result.push('u');
-            return;
-        }
-        // Incomplete \u escape (not enough chars)
-        self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-        result.push('\\');
-        result.push('u');
-    }
-
-    /// Scan a template literal (simplified).
-    fn scan_template_literal(&mut self) {
-        self.pos += 1; // Skip backtick
-        let mut result = String::new();
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKTICK {
-                self.pos += 1;
-                self.token_value = result;
-                self.token = SyntaxKind::NoSubstitutionTemplateLiteral;
-                return;
-            }
-            if ch == CharacterCodes::DOLLAR
-                && self.char_code_at(self.pos + 1) == Some(CharacterCodes::OPEN_BRACE)
-            {
-                self.pos += 2;
-                self.token_value = result;
-                self.token = SyntaxKind::TemplateHead;
-                return;
-            }
-            if ch == CharacterCodes::BACKSLASH {
-                // Scan escaped character after the backslash.
-                self.pos += 1;
-                let escaped = self.scan_template_escape_sequence();
-                result.push_str(&escaped);
-            } else {
-                if ch == CharacterCodes::LINE_FEED || ch == CharacterCodes::CARRIAGE_RETURN {
-                    self.token_flags |= TokenFlags::PrecedingLineBreak as u32;
-                }
-                if let Some(c) = char::from_u32(ch) {
-                    result.push(c);
-                }
-                self.pos += self.char_len_at(self.pos); // Advance by character byte length
-            }
-        }
-
-        self.token_flags |= TokenFlags::Unterminated as u32;
-        self.token_value = result;
-        self.token = SyntaxKind::NoSubstitutionTemplateLiteral;
-    }
-
-    /// Scan a number literal (simplified).
-    fn scan_number(&mut self) {
-        let start = self.pos;
-
-        // Check for hex, octal, binary
-        if self.char_code_unchecked(self.pos) == CharacterCodes::_0 {
-            let next = self.char_code_at(self.pos + 1).unwrap_or(0);
-            if self.scan_prefixed_number(start, next) {
-                return;
-            }
-
-            // After leading 0, scan all consecutive digits and check if all are octal.
-            // This matches tsc's scanDigits(): scan 0-9, return whether all were 0-7.
-            if is_digit(next) && self.scan_legacy_octal_number(start) {
-                return;
-            }
-
-            // Numeric separator immediately after a leading zero (e.g. `0_0`,
-            // `0_1.5`, `0__0`) — tsc treats this as a legacy-octal-style
-            // leading zero and rejects the separator with TS6188. The
-            // integer part of a decimal literal that starts with a `0`
-            // followed by `_` and any continuation (digit OR another `_`)
-            // is not a valid form. Emit the diagnostic at the `_` position
-            // and fall through to `scan_decimal_number` so the rest of the
-            // literal still produces the expected NumericLiteral token.
-            //
-            // The `_<digit>` arm catches `0_0`. The `_<_>` arm catches
-            // `0__0` (regression: `parser.numericSeparators.decmialNegative.ts`
-            // files 18/31/44, where a consecutive-separator run after the
-            // leading zero ends in a digit; the inner loop emits TS6189 at
-            // the second `_`, but the leading-zero TS6188 at the first `_`
-            // was missing without this check).
-            if next == CharacterCodes::UNDERSCORE
-                && self
-                    .char_code_at(self.pos + 2)
-                    .is_some_and(|c| is_digit(c) || c == CharacterCodes::UNDERSCORE)
-            {
-                self.scanner_diagnostics.push(ScannerDiagnostic {
-                    pos: self.pos + 1,
-                    length: 1,
-                    args: Vec::new(),
-                    message: diagnostic_messages::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE,
-                    code: diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE,
-                });
-                self.token_flags |= TokenFlags::ContainsInvalidSeparator as u32;
-                if self.token_invalid_separator_pos.is_none() {
-                    self.token_invalid_separator_pos = Some(self.pos + 1);
-                    self.token_invalid_separator_is_consecutive = false;
-                }
-            }
-        }
-
-        // Decimal number
-        self.scan_decimal_number(start);
-    }
-
-    fn scan_prefixed_number(&mut self, start: usize, next: u32) -> bool {
-        match next {
-            CharacterCodes::LOWER_X | CharacterCodes::UPPER_X => {
-                self.scan_integer_base_literal(start, is_hex_digit, TokenFlags::HexSpecifier);
-                true
-            }
-            CharacterCodes::LOWER_B | CharacterCodes::UPPER_B => {
-                self.scan_integer_base_literal(start, is_binary_digit, TokenFlags::BinarySpecifier);
-                true
-            }
-            CharacterCodes::LOWER_O | CharacterCodes::UPPER_O => {
-                self.scan_integer_base_literal(start, is_octal_digit, TokenFlags::OctalSpecifier);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    fn scan_integer_base_literal(
-        &mut self,
-        start: usize,
-        is_valid_digit: fn(u32) -> bool,
-        specifier_flag: TokenFlags,
-    ) {
-        self.pos += 2;
-        self.token_flags |= specifier_flag as u32;
-        let saw_digit = self.scan_digits_with_separators(is_valid_digit);
-
-        // tsc parity: if the prefixed integer literal has no valid digits
-        // (e.g. `0xn`, `0bn`, `0on`, `0x`, `0x_`), emit a zero-width
-        // "<base> digit expected" diagnostic at the position right after the
-        // base prefix. Mirrors `scanner.ts`'s `if (!tokenValue) error(...)`
-        // ladder in `scanIntegerBaseLiteral`/`case CharacterCodes._0`.
-        if !saw_digit {
-            let flag = specifier_flag as u32;
-            let (message, code) = if flag == TokenFlags::HexSpecifier as u32 {
-                (
-                    diagnostic_messages::HEXADECIMAL_DIGIT_EXPECTED,
-                    diagnostic_codes::HEXADECIMAL_DIGIT_EXPECTED,
-                )
-            } else if flag == TokenFlags::BinarySpecifier as u32 {
-                (
-                    diagnostic_messages::BINARY_DIGIT_EXPECTED,
-                    diagnostic_codes::BINARY_DIGIT_EXPECTED,
-                )
-            } else {
-                (
-                    diagnostic_messages::OCTAL_DIGIT_EXPECTED,
-                    diagnostic_codes::OCTAL_DIGIT_EXPECTED,
-                )
-            };
-            self.scanner_diagnostics.push(ScannerDiagnostic {
-                pos: self.pos,
-                length: 0,
-                args: Vec::new(),
-                message,
-                code,
-            });
-        }
-
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::LOWER_N {
-            self.pos += 1;
-            self.token_value = self.substring(start, self.pos);
-            self.check_for_identifier_start_after_numeric_literal(
-                start, /*is_scientific*/ false,
-            );
-            self.token = SyntaxKind::BigIntLiteral;
-            return;
-        }
-
-        self.set_numeric_token_value(start);
-        self.check_for_identifier_start_after_numeric_literal(start, /*is_scientific*/ false);
-        self.token = SyntaxKind::NumericLiteral;
-    }
-
-    fn scan_legacy_octal_number(&mut self, start: usize) -> bool {
-        let mut all_octal = true;
-        let digit_start = self.pos + 1; // skip the leading 0
-        let mut scan_pos = digit_start;
-        while scan_pos < self.end && is_digit(self.char_code_unchecked(scan_pos)) {
-            if !is_octal_digit(self.char_code_unchecked(scan_pos)) {
-                all_octal = false;
-            }
-            scan_pos += 1;
-        }
-        if all_octal && scan_pos > digit_start {
-            self.pos = scan_pos;
-            self.token_flags |= TokenFlags::Octal as u32;
-            self.set_numeric_token_value(start);
-            self.token = SyntaxKind::NumericLiteral;
-            true
-        } else {
-            self.token_flags |= TokenFlags::ContainsLeadingZero as u32;
-            false
-        }
-    }
-
-    fn set_numeric_token_value(&mut self, start: usize) {
-        if (self.token_flags & TokenFlags::ContainsSeparator as u32) != 0 {
-            self.token_value = self.substring(start, self.pos);
-        } else {
-            self.token_value.clear();
-        }
-    }
-
-    fn scan_decimal_number(&mut self, start: usize) {
-        self.scan_digits_with_separators(is_digit);
-        let mut has_decimal_point = false;
-
-        // Decimal point
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::DOT {
-            has_decimal_point = true;
-            self.pos += 1;
-            self.scan_digits_with_separators(is_digit);
-        }
-
-        // Exponent
-        let mut has_exponent = false;
-        if self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::LOWER_E || ch == CharacterCodes::UPPER_E {
-                has_exponent = true;
-                self.pos += 1;
-                self.token_flags |= TokenFlags::Scientific as u32;
-                if self.pos < self.end {
-                    let sign = self.char_code_unchecked(self.pos);
-                    if sign == CharacterCodes::PLUS || sign == CharacterCodes::MINUS {
-                        self.pos += 1;
-                    }
-                }
-                // Mirror tsc's `scanNumber` exponent branch: if the exponent has
-                // no digits (e.g. `1e`, `1e+`, `1ee`, `3en`), emit TS1124
-                // "Digit expected" at the current position. Emitting here (before
-                // `check_for_identifier_start_after_numeric_literal`) lets us
-                // mirror tsc's `parseErrorAtPosition` same-start dedup: a later
-                // TS1351 at the same position is suppressed by the helper below.
-                let saw_exp_digit = self.scan_digits_with_separators(is_digit);
-                if !saw_exp_digit {
-                    self.scanner_diagnostics.push(ScannerDiagnostic {
-                        pos: self.pos,
-                        length: 0,
-                        args: Vec::new(),
-                        message: diagnostic_messages::DIGIT_EXPECTED,
-                        code: diagnostic_codes::DIGIT_EXPECTED,
-                    });
-                }
-            }
-        }
-
-        // BigInt suffix
-        if !has_decimal_point
-            && !has_exponent
-            && self.pos < self.end
-            && self.char_code_unchecked(self.pos) == CharacterCodes::LOWER_N
-        {
-            self.pos += 1;
-            self.token_value = self.substring(start, self.pos);
-            self.check_for_identifier_start_after_numeric_literal(
-                start, /*is_scientific*/ false,
-            );
-            self.token = SyntaxKind::BigIntLiteral;
-            return;
-        }
-
-        // OPTIMIZATION: Only allocate token_value if number contains separators
-        // Plain numbers (no underscores) can use source slice via get_token_value_ref()
-        self.set_numeric_token_value(start);
-        // When the recovery branch consumed an `n` BigInt suffix on a non-integer
-        // (e.g. `3en`), `check_for_identifier_start_after_numeric_literal` returns
-        // `true` and `self.pos` is now past the `n`. Re-capture `token_value` from
-        // the full consumed span so emit preserves the source spelling, matching
-        // tsc which prints `3en[null];` for `3en[null]` rather than `3e[null];`.
-        if self.check_for_identifier_start_after_numeric_literal(start, has_exponent) {
-            self.token_value = self.substring(start, self.pos);
-        }
-        self.token = SyntaxKind::NumericLiteral;
-    }
-
-    fn check_for_identifier_start_after_numeric_literal(
-        &mut self,
-        numeric_start: usize,
-        is_scientific: bool,
-    ) -> bool {
-        if self.pos >= self.end {
-            return false;
-        }
-
-        // Only check the raw character code, not unicode escapes.
-        // TSC uses `codePointAt(text, pos)` here which reads the literal char,
-        // so `\u005F` (backslash) is NOT treated as an identifier start.
-        let starts_identifier = self.is_identifier_start(self.char_code_unchecked(self.pos));
-
-        if !starts_identifier {
-            return false;
-        }
-
-        let identifier_start = self.pos;
-        self.scan_identifier_parts_after_numeric_literal();
-        let identifier_end = self.pos;
-        let identifier_text = &self.source[identifier_start..identifier_end];
-
-        if identifier_text == "n" {
-            self.scanner_diagnostics.push(ScannerDiagnostic {
-                pos: numeric_start,
-                length: identifier_end - numeric_start,
-                args: Vec::new(),
-                message: if is_scientific {
-                    diagnostic_messages::A_BIGINT_LITERAL_CANNOT_USE_EXPONENTIAL_NOTATION
-                } else {
-                    diagnostic_messages::A_BIGINT_LITERAL_MUST_BE_AN_INTEGER
-                },
-                code: if is_scientific {
-                    diagnostic_codes::A_BIGINT_LITERAL_CANNOT_USE_EXPONENTIAL_NOTATION
-                } else {
-                    diagnostic_codes::A_BIGINT_LITERAL_MUST_BE_AN_INTEGER
-                },
-            });
-            true
-        } else {
-            // Mirror tsc's `parseErrorAtPosition` same-start dedup: if a prior
-            // scanner diagnostic was already pushed at `identifier_start`
-            // (e.g. TS1124 "Digit expected" emitted by the empty-exponent
-            // branch in `scan_decimal_number` for `1ee`/`123ee`), tsc's parser
-            // would suppress this TS1351 because its `lastError.start` matches.
-            // We mirror that suppression here so the merged diagnostics match
-            // tsc fingerprint-for-fingerprint.
-            let already_diag_at_pos = self
-                .scanner_diagnostics
-                .last()
-                .is_some_and(|d| d.pos == identifier_start);
-            if !already_diag_at_pos {
-                self.scanner_diagnostics.push(ScannerDiagnostic {
-                    pos: identifier_start,
-                    length: identifier_end - identifier_start,
-                    message: diagnostic_messages::AN_IDENTIFIER_OR_KEYWORD_CANNOT_IMMEDIATELY_FOLLOW_A_NUMERIC_LITERAL,
-                    code: diagnostic_codes::AN_IDENTIFIER_OR_KEYWORD_CANNOT_IMMEDIATELY_FOLLOW_A_NUMERIC_LITERAL,
-                    args: Vec::new(),
-                });
-            }
-            self.pos = identifier_start;
-            false
-        }
-    }
-
-    fn scan_identifier_parts_after_numeric_literal(&mut self) {
-        if self.pos >= self.end {
-            return;
-        }
-
-        if self.char_code_unchecked(self.pos) == CharacterCodes::BACKSLASH {
-            if let Some(code_point) = self.peek_unicode_escape() {
-                if self.is_identifier_start(code_point) {
-                    let _ = self.scan_unicode_escape_value();
-                } else {
-                    return;
-                }
-            } else {
-                return;
-            }
-        } else if self.is_identifier_start(self.char_code_unchecked(self.pos)) {
-            self.pos += self.char_len_at(self.pos);
-        } else {
-            return;
-        }
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    let _ = self.scan_unicode_escape_value();
-                    continue;
-                }
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            self.pos += self.char_len_at(self.pos);
-        }
-    }
-
-    /// Returns `true` if at least one valid digit was consumed. Mirrors tsc's
-    /// `scanHexDigits`/`scanBinaryOrOctalDigits` "did we read anything" signal,
-    /// so callers like `scan_integer_base_literal` can emit
-    /// `Hexadecimal/Binary/Octal digit expected` for empty prefixed literals.
-    /// Also emits one diagnostic per invalid numeric separator as we walk the
-    /// digit run, distinguishing TS6189 ("multiple consecutive") from TS6188
-    /// ("not allowed here") by tracking whether the immediately preceding
-    /// token was a *valid* separator.
-    fn scan_digits_with_separators(&mut self, is_valid_digit: fn(u32) -> bool) -> bool {
-        let scan_start = self.pos;
-        let mut saw_digit = false;
-        let mut allow_separator = false;
-        let mut is_previous_separator = false;
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::UNDERSCORE {
-                self.token_flags |= TokenFlags::ContainsSeparator as u32;
-                if allow_separator {
-                    allow_separator = false;
-                    is_previous_separator = true;
-                } else if is_previous_separator {
-                    self.push_invalid_numeric_separator(self.pos, /*consecutive*/ true);
-                } else {
-                    self.push_invalid_numeric_separator(self.pos, /*consecutive*/ false);
-                }
-                self.pos += 1;
-                continue;
-            }
-            if is_valid_digit(ch) {
-                saw_digit = true;
-                allow_separator = true;
-                is_previous_separator = false;
-                self.pos += 1;
-                continue;
-            }
-            break;
-        }
-
-        // Trailing underscore — tsc fires Numeric_separators_are_not_allowed_here
-        // unconditionally if the byte before pos is `_`. The scanner-level
-        // dedup in `push_invalid_numeric_separator` collapses the same-position
-        // pair (TS6189 + TS6188) that arises when the inner loop already emitted
-        // for that same byte. Bound the look-back at `scan_start` so a zero-iter
-        // call (e.g., a `.5` decimal where the first byte is `.`) cannot inspect
-        // a byte from a preceding token like the `_` in `_.5`.
-        if self.pos > scan_start
-            && self.char_code_unchecked(self.pos - 1) == CharacterCodes::UNDERSCORE
-        {
-            self.push_invalid_numeric_separator(self.pos - 1, /*consecutive*/ false);
-        }
-
-        saw_digit
-    }
-
-    /// Push a numeric-separator diagnostic, applying the same "skip if last
-    /// diagnostic shares this start position" rule that tsc's
-    /// `parseErrorAtPosition` enforces (only the first of a same-start run
-    /// survives). Also records the first invalid separator position for the
-    /// parser's recovery path (e.g. TS2304 emission for `0_X0101`).
-    fn push_invalid_numeric_separator(&mut self, pos: usize, consecutive: bool) {
-        self.token_flags |= TokenFlags::ContainsInvalidSeparator as u32;
-        if self.token_invalid_separator_pos.is_none() {
-            self.token_invalid_separator_pos = Some(pos);
-            self.token_invalid_separator_is_consecutive = consecutive;
-        }
-        if let Some(last) = self.scanner_diagnostics.last()
-            && last.pos == pos
-        {
-            return;
-        }
-        let (message, code) = if consecutive {
-            (
-                diagnostic_messages::MULTIPLE_CONSECUTIVE_NUMERIC_SEPARATORS_ARE_NOT_PERMITTED,
-                diagnostic_codes::MULTIPLE_CONSECUTIVE_NUMERIC_SEPARATORS_ARE_NOT_PERMITTED,
-            )
-        } else {
-            (
-                diagnostic_messages::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE,
-                diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE,
-            )
-        };
-        self.scanner_diagnostics.push(ScannerDiagnostic {
-            pos,
-            length: 1,
-            message,
-            code,
-            args: Vec::new(),
-        });
-    }
-
-    fn push_invalid_character(&mut self, pos: usize) {
-        if let Some(last) = self.scanner_diagnostics.last()
-            && last.pos == pos
-        {
-            return;
-        }
-        self.scanner_diagnostics.push(ScannerDiagnostic {
-            pos,
-            length: 1,
-            message: diagnostic_messages::INVALID_CHARACTER,
-            code: diagnostic_codes::INVALID_CHARACTER,
-            args: Vec::new(),
-        });
-    }
-
-    /// Scan an identifier.
-    /// ZERO-ALLOCATION: Identifiers are interned, returning an Atom (u32) for O(1) comparison.
-    /// When a unicode escape is encountered mid-identifier, switches to allocation mode.
-    fn scan_identifier(&mut self) {
-        let start = self.pos;
-        // Advance past first character (may be multi-byte)
-        self.pos += self.char_len_at(self.pos);
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                // Check if this is a unicode escape that produces an identifier part
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    // Switch to allocation mode and continue scanning with escapes
-                    self.continue_identifier_with_escapes(start);
-                    return;
-                }
-                if self.peek_unicode_escape().is_some() {
-                    self.push_invalid_character(self.pos);
-                }
-                // Invalid escape or not an identifier part - stop here
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            self.pos += self.char_len_at(self.pos); // Handle multi-byte UTF-8
-        }
-
-        // Get slice reference instead of allocating new String
-        let text_slice = &self.source[start..self.pos];
-
-        // Check if it's a keyword first (common keywords are pre-interned)
-        self.token = crate::text_to_keyword(text_slice).unwrap_or(SyntaxKind::Identifier);
-
-        // Intern the identifier for O(1) comparison (reuses existing interned string)
-        self.token_atom = self.interner.intern(text_slice);
-
-        // ZERO-ALLOCATION: Don't store token_value for identifiers.
-        // get_token_value_ref() will resolve from token_atom or fall back to source slice.
-        self.token_value.clear();
-    }
-
-    /// Continue scanning an identifier that has a unicode escape mid-identifier.
-    /// Called when `scan_identifier()` encounters a valid unicode escape.
-    /// This switches to allocation mode since escapes require building a String.
-    fn continue_identifier_with_escapes(&mut self, start: usize) {
-        // Copy the already-scanned part into a String
-        let mut result = String::from(&self.source[start..self.pos]);
-
-        // Continue scanning identifier parts, handling unicode escapes
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                // Check for unicode escape
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    // Consume the escape and add the character
-                    if let Some(c) = char::from_u32(self.scan_unicode_escape_value().unwrap_or(0)) {
-                        result.push(c);
-                    }
-                    continue;
-                }
-                if self.peek_unicode_escape().is_some() {
-                    self.push_invalid_character(self.pos);
-                }
-                // Invalid escape or not an identifier part - stop here
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            if let Some(c) = char::from_u32(ch) {
-                result.push(c);
-            }
-            self.pos += self.char_len_at(self.pos);
-        }
-
-        self.token = crate::text_to_keyword(&result).unwrap_or(SyntaxKind::Identifier);
-        self.token_atom = self.interner.intern(&result);
-        self.token_value.clear();
-        self.token_flags |= TokenFlags::UnicodeEscape as u32;
-    }
-
-    /// Peek at a unicode escape sequence in identifier text without advancing
-    /// the position. Returns the code point if the escape is valid
-    /// (`\uXXXX`, or `\u{XXXXX}` for a valid Unicode scalar), None otherwise.
-    pub(crate) fn peek_unicode_escape(&self) -> Option<u32> {
-        // Must start with \u
-        if self.pos + 1 >= self.end {
-            return None;
-        }
-        let bytes = self.source.as_bytes();
-        if bytes.get(self.pos + 1).copied() != Some(b'u') {
-            return None;
-        }
-        // \u{XXXXX} form
-        if bytes.get(self.pos + 2).copied() == Some(b'{') {
-            let start = self.pos + 3;
-            let mut end = start;
-            while end < self.end && bytes.get(end).is_some_and(u8::is_ascii_hexdigit) {
-                end += 1;
-            }
-            if end == start || bytes.get(end).copied() != Some(b'}') {
-                return None;
-            }
-            let hex = &self.source[start..end];
-            u32::from_str_radix(hex, 16)
-                .ok()
-                .filter(|&cp| char::from_u32(cp).is_some())
-        } else {
-            // \uXXXX form (exactly 4 hex digits)
-            if self.pos + 5 >= self.end {
-                return None;
-            }
-            let hex = &self.source[self.pos + 2..self.pos + 6];
-            if hex.len() == 4 && hex.bytes().all(|b| b.is_ascii_hexdigit()) {
-                u32::from_str_radix(hex, 16).ok()
-            } else {
-                None
-            }
-        }
-    }
-
-    /// Scan an identifier that starts with a unicode escape sequence (\uXXXX).
-    fn scan_identifier_with_escapes(&mut self) {
-        let mut result = String::new();
-
-        // Process initial unicode escape
-        if let Some(ch) = self.scan_unicode_escape_value()
-            && let Some(c) = char::from_u32(ch)
-        {
-            result.push(c);
-        }
-
-        // Continue scanning identifier parts
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                // Another unicode escape in identifier
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    if let Some(c) = char::from_u32(self.scan_unicode_escape_value().unwrap_or(0)) {
-                        result.push(c);
-                    }
-                    continue;
-                }
-                if self.peek_unicode_escape().is_some() {
-                    self.push_invalid_character(self.pos);
-                }
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            if let Some(c) = char::from_u32(ch) {
-                result.push(c);
-            }
-            self.pos += self.char_len_at(self.pos);
-        }
-
-        self.token = crate::text_to_keyword(&result).unwrap_or(SyntaxKind::Identifier);
-        self.token_atom = self.interner.intern(&result);
-        self.token_value.clear();
-        self.token_flags |= TokenFlags::UnicodeEscape as u32;
-    }
-
-    /// Scan continuation of a private identifier that starts with a regular char.
-    /// Handles unicode escapes in the continuation (e.g., `#x\u0078`).
-    /// Returns true if any unicode escapes were found and `token_value` was set.
-    pub(crate) fn scan_private_identifier_rest(&mut self) -> bool {
-        // First, scan regular identifier parts
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                // Found a unicode escape in the continuation
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    // Build the decoded value from the start
-                    let prefix = self.substring(self.token_start, self.pos);
-                    let mut result = prefix;
-                    // Consume the escape
-                    if let Some(cp) = self.scan_unicode_escape_value()
-                        && let Some(c) = char::from_u32(cp)
-                    {
-                        result.push(c);
-                    }
-                    // Continue scanning the rest
-                    while self.pos < self.end {
-                        let ch2 = self.char_code_unchecked(self.pos);
-                        if ch2 == CharacterCodes::BACKSLASH {
-                            if let Some(cp2) = self.peek_unicode_escape()
-                                && self.is_identifier_part(cp2)
-                            {
-                                if let Some(cp2) = self.scan_unicode_escape_value()
-                                    && let Some(c) = char::from_u32(cp2)
-                                {
-                                    result.push(c);
-                                }
-                                continue;
-                            }
-                            if self.peek_unicode_escape().is_some() {
-                                self.push_invalid_character(self.pos);
-                            }
-                            break;
-                        }
-                        if !self.is_identifier_part(ch2) {
-                            break;
-                        }
-                        if let Some(c) = char::from_u32(ch2) {
-                            result.push(c);
-                        }
-                        self.pos += self.char_len_at(self.pos);
-                    }
-                    self.token_value = result;
-                    self.token_flags |= TokenFlags::UnicodeEscape as u32;
-                    return true;
-                }
-                if self.peek_unicode_escape().is_some() {
-                    self.push_invalid_character(self.pos);
-                }
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            self.pos += self.char_len_at(self.pos);
-        }
-        false
-    }
-
-    /// Scan a private identifier that starts with a unicode escape: `#\u0078`.
-    pub(crate) fn scan_private_identifier_with_escapes(&mut self) {
-        let mut result = String::from("#");
-
-        // Process initial unicode escape
-        if let Some(ch) = self.scan_unicode_escape_value()
-            && let Some(c) = char::from_u32(ch)
-        {
-            result.push(c);
-        }
-
-        // Continue scanning identifier parts
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == CharacterCodes::BACKSLASH {
-                // Another unicode escape in identifier
-                if let Some(code_point) = self.peek_unicode_escape()
-                    && self.is_identifier_part(code_point)
-                {
-                    if let Some(c) = char::from_u32(self.scan_unicode_escape_value().unwrap_or(0)) {
-                        result.push(c);
-                    }
-                    continue;
-                }
-                if self.peek_unicode_escape().is_some() {
-                    self.push_invalid_character(self.pos);
-                }
-                break;
-            }
-            if !self.is_identifier_part(ch) {
-                break;
-            }
-            if let Some(c) = char::from_u32(ch) {
-                result.push(c);
-            }
-            self.pos += self.char_len_at(self.pos);
-        }
-
-        self.token = SyntaxKind::PrivateIdentifier;
-        self.token_atom = self.interner.intern(&result);
-        self.token_value = result;
-        self.token_flags |= TokenFlags::UnicodeEscape as u32;
-    }
-
-    /// Consume a unicode escape sequence and return its code point.
-    /// Advances self.pos past the escape.
-    fn scan_unicode_escape_value(&mut self) -> Option<u32> {
-        // Skip the backslash
-        self.pos += 1;
-        if self.pos >= self.end || self.source.as_bytes()[self.pos] != b'u' {
-            return None;
-        }
-        self.pos += 1; // Skip 'u'
-
-        if self.pos < self.end && self.source.as_bytes()[self.pos] == b'{' {
-            // \u{XXXXX} form
-            self.pos += 1;
-            let start = self.pos;
-            while self.pos < self.end
-                && self
-                    .source
-                    .as_bytes()
-                    .get(self.pos)
-                    .is_some_and(u8::is_ascii_hexdigit)
-            {
-                self.pos += 1;
-            }
-            let result = u32::from_str_radix(&self.source[start..self.pos], 16).ok();
-            if self.pos < self.end && self.source.as_bytes()[self.pos] == b'}' {
-                self.pos += 1;
-            }
-            result
-        } else {
-            // \uXXXX form (exactly 4 hex digits)
-            if self.pos + 4 > self.end {
-                return None;
-            }
-            let hex = &self.source[self.pos..self.pos + 4];
-            if hex.bytes().all(|b| b.is_ascii_hexdigit()) {
-                self.pos += 4;
-                u32::from_str_radix(hex, 16).ok()
-            } else {
-                None
-            }
-        }
-    }
-
-    // =========================================================================
-    // Rescan methods - for context-sensitive parsing
-    //
-    // Many rescan modes live in the `rescan` sibling module so the mode-shifting
-    // surface is isolated from the main `scan()` loop. Template, slash, and JSX
-    // rescans remain here because they share many private scanning helpers with
-    // the main scan path.
-    // =========================================================================
-
-    /// Re-scan the current `/` or `/=` token as a regex literal.
-    /// This is used by the parser when it determines the context requires a regex.
-    #[wasm_bindgen(js_name = reScanSlashToken)]
-    pub fn re_scan_slash_token(&mut self) -> SyntaxKind {
-        if self.token == SyntaxKind::SlashToken || self.token == SyntaxKind::SlashEqualsToken {
-            // Start scanning from after the initial /
-            let start_of_regex_body = self.token_start + 1;
-            self.pos = start_of_regex_body;
-            let mut in_escape = false;
-            let mut in_character_class = false;
-
-            // Scan until we find the closing /
-            while self.pos < self.end {
-                let ch = self.char_code_unchecked(self.pos);
-
-                // Unterminated regex if we hit a newline
-                if is_line_break(ch) {
-                    self.token_flags |= TokenFlags::Unterminated as u32;
-                    break;
-                }
-
-                if in_escape {
-                    // After backslash, just consume the next character
-                    in_escape = false;
-                } else if ch == CharacterCodes::SLASH && !in_character_class {
-                    // Found the closing /
-                    break;
-                } else if ch == CharacterCodes::OPEN_BRACKET {
-                    in_character_class = true;
-                } else if ch == CharacterCodes::BACKSLASH {
-                    in_escape = true;
-                } else if ch == CharacterCodes::CLOSE_BRACKET {
-                    in_character_class = false;
-                }
-                // Use char_len_at to properly advance past multi-byte UTF-8 characters
-                self.pos += self.char_len_at(self.pos);
-            }
-
-            // If we reached EOF without finding closing /, mark as unterminated
-            if self.pos >= self.end && (self.token_flags & TokenFlags::Unterminated as u32) == 0 {
-                self.token_flags |= TokenFlags::Unterminated as u32;
-            }
-
-            if (self.token_flags & TokenFlags::Unterminated as u32) == 0 {
-                // Consume the closing /
-                self.pos += 1;
-
-                // Scan and validate regex flags (g, i, m, s, u, v, y, d)
-                // Track seen flags as a bitmask for duplicate detection
-                let mut seen_flags: u8 = 0;
-                let mut has_u = false;
-                let mut has_v = false;
-
-                while self.pos < self.end {
-                    let ch = self.char_code_unchecked(self.pos);
-                    if !is_regex_flag(ch) && !is_identifier_part(ch) {
-                        break;
-                    }
-
-                    // Check for valid flags and detect errors
-                    let flag_bit = match ch {
-                        CharacterCodes::LOWER_G => Some(0),
-                        CharacterCodes::LOWER_I => Some(1),
-                        CharacterCodes::LOWER_M => Some(2),
-                        CharacterCodes::LOWER_S => Some(3),
-                        CharacterCodes::LOWER_U => {
-                            has_u = true;
-                            Some(4)
-                        }
-                        CharacterCodes::LOWER_V => {
-                            has_v = true;
-                            Some(5)
-                        }
-                        CharacterCodes::LOWER_Y => Some(6),
-                        CharacterCodes::LOWER_D => Some(7),
-                        _ => None,
-                    };
-
-                    if let Some(bit) = flag_bit {
-                        let mask = 1 << bit;
-                        if seen_flags & mask != 0 {
-                            // Duplicate flag - emit error for each duplicate
-                            self.regex_flag_errors.push(RegexFlagError {
-                                kind: RegexFlagErrorKind::Duplicate,
-                                pos: self.pos,
-                            });
-                        }
-                        seen_flags |= mask;
-                    } else if is_identifier_part(ch) {
-                        // Invalid flag character (identifier char but not a valid flag)
-                        self.regex_flag_errors.push(RegexFlagError {
-                            kind: RegexFlagErrorKind::InvalidFlag,
-                            pos: self.pos,
-                        });
-                    }
-
-                    // Use char_len_at for proper UTF-8 handling (handles non-ASCII flags)
-                    self.pos += self.char_len_at(self.pos);
-                }
-
-                // Check for incompatible u and v flags
-                if has_u && has_v {
-                    // Emit error at the end of flags (similar to TypeScript)
-                    self.regex_flag_errors.push(RegexFlagError {
-                        kind: RegexFlagErrorKind::IncompatibleFlags,
-                        pos: self.pos,
-                    });
-                }
-            }
-
-            self.token_value = self.substring(self.token_start, self.pos);
-            self.token = SyntaxKind::RegularExpressionLiteral;
-        }
-        self.token
-    }
-
-    /// Re-scan the current `}` token as the continuation of a template literal.
-    /// Called by the parser when it determines that a `}` is closing a template expression.
-    ///
-    /// # Arguments
-    /// * `is_tagged_template` - If true, invalid escape sequences should not report errors
-    ///   (tagged templates can have invalid escapes that get passed to the tag function as raw).
-    ///   For now, we don't report errors anyway, so this parameter affects nothing.
-    #[wasm_bindgen(js_name = reScanTemplateToken)]
-    pub fn re_scan_template_token(&mut self, _is_tagged_template: bool) -> SyntaxKind {
-        // Reset position to token start and scan the template continuation
-        // Make sure token_start is within bounds
-        if self.token_start >= self.end {
-            self.token = SyntaxKind::EndOfFileToken;
-            return self.token;
-        }
-        self.pos = self.token_start;
-        self.token = self.scan_template_and_set_token_value(false);
-        self.token
-    }
-
-    /// Re-scan template head or no-substitution template.
-    /// Used when the parser needs to rescan the start of a template.
-    #[wasm_bindgen(js_name = reScanTemplateHeadOrNoSubstitutionTemplate)]
-    pub fn re_scan_template_head_or_no_substitution_template(&mut self) -> SyntaxKind {
-        self.pos = self.token_start;
-        self.token = self.scan_template_and_set_token_value(true);
-        self.token
-    }
-
-    /// Internal helper to scan a template literal part and set the token value.
-    ///
-    /// # Arguments
-    /// * `started_with_backtick` - true if this is the start of a template (head or no-substitution),
-    ///   false if this is a continuation after a `}` (middle or tail).
-    fn scan_template_and_set_token_value(&mut self, started_with_backtick: bool) -> SyntaxKind {
-        // Move past the opening character (backtick for head, } for middle/tail)
-        // Safety check: ensure we don't move past the end
-        if self.pos >= self.end {
-            self.token_flags |= TokenFlags::Unterminated as u32;
-            self.token_value = String::new();
-            return if started_with_backtick {
-                SyntaxKind::NoSubstitutionTemplateLiteral
-            } else {
-                SyntaxKind::TemplateTail
-            };
-        }
-        self.pos += 1;
-        let mut start = self.pos;
-        let mut contents = String::new();
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-
-            // End of template: backtick
-            if ch == CharacterCodes::BACKTICK {
-                contents.push_str(&self.substring(start, self.pos));
-                self.pos += 1;
-                self.token_value = contents;
-                return if started_with_backtick {
-                    SyntaxKind::NoSubstitutionTemplateLiteral
-                } else {
-                    SyntaxKind::TemplateTail
-                };
-            }
-
-            // Template expression: ${
-            if ch == CharacterCodes::DOLLAR
-                && self.pos + 1 < self.end
-                && self.char_code_unchecked(self.pos + 1) == CharacterCodes::OPEN_BRACE
-            {
-                contents.push_str(&self.substring(start, self.pos));
-                self.pos += 2;
-                self.token_value = contents;
-                return if started_with_backtick {
-                    SyntaxKind::TemplateHead
-                } else {
-                    SyntaxKind::TemplateMiddle
-                };
-            }
-
-            // Escape sequence
-            if ch == CharacterCodes::BACKSLASH {
-                contents.push_str(&self.substring(start, self.pos));
-                self.pos += 1; // Advance past the backslash
-                let escaped = self.scan_template_escape_sequence();
-                contents.push_str(&escaped);
-                // Reset start to current position after the escape
-                start = self.pos;
-                continue;
-            }
-
-            // CR normalization (CR or CRLF -> LF)
-            if ch == CharacterCodes::CARRIAGE_RETURN {
-                contents.push_str(&self.substring(start, self.pos));
-                self.pos += 1;
-                if self.pos < self.end
-                    && self.char_code_unchecked(self.pos) == CharacterCodes::LINE_FEED
-                {
-                    self.pos += 1;
-                }
-                contents.push('\n');
-                // Reset start to current position after the CR
-                start = self.pos;
-                continue;
-            }
-
-            // Advance by full UTF-8 codepoint width so multi-byte chars (e.g. µ) don't
-            // move the scanner into a non-char-boundary byte index.
-            self.pos += self.char_len_at(self.pos);
-        }
-
-        // Unterminated template
-        contents.push_str(&self.substring(start, self.pos));
-        self.token_flags |= TokenFlags::Unterminated as u32;
-        self.token_value = contents;
-        if started_with_backtick {
-            SyntaxKind::NoSubstitutionTemplateLiteral
-        } else {
-            SyntaxKind::TemplateTail
-        }
-    }
-
-    /// Scan an escape sequence in a template literal.
-    /// Returns the resulting string and advances self.pos.
-    ///
-    /// `self.pos` is expected to point at the escaped character (just after `\`).
-    fn scan_template_escape_sequence(&mut self) -> String {
-        if self.pos >= self.end {
-            return String::from("\\");
-        }
-
-        let ch = self.char_code_unchecked(self.pos);
-        // Use char_len_at for proper UTF-8 handling of multi-byte chars
-        let ch_len = self.char_len_at(self.pos);
-        self.pos += ch_len;
-
-        match ch {
-            CharacterCodes::_0 => self.scan_template_escape_digit_zero(),
-            CharacterCodes::_1
-            | CharacterCodes::_2
-            | CharacterCodes::_3
-            | CharacterCodes::_4
-            | CharacterCodes::_5
-            | CharacterCodes::_6
-            | CharacterCodes::_7
-            | CharacterCodes::_8
-            | CharacterCodes::_9 => self.scan_template_escape_octal_digit(ch),
-            CharacterCodes::LOWER_N => String::from("\n"),
-            CharacterCodes::LOWER_R => String::from("\r"),
-            CharacterCodes::LOWER_T => String::from("\t"),
-            CharacterCodes::LOWER_V => String::from("\x0B"),
-            CharacterCodes::LOWER_B => String::from("\x08"),
-            CharacterCodes::LOWER_F => String::from("\x0C"),
-            CharacterCodes::SINGLE_QUOTE => String::from("'"),
-            CharacterCodes::DOUBLE_QUOTE => String::from("\""),
-            CharacterCodes::BACKTICK => String::from("`"),
-            CharacterCodes::BACKSLASH => String::from("\\"),
-            CharacterCodes::DOLLAR => String::from("$"),
-            CharacterCodes::LINE_FEED
-            | CharacterCodes::LINE_SEPARATOR
-            | CharacterCodes::PARAGRAPH_SEPARATOR => String::new(),
-            CharacterCodes::CARRIAGE_RETURN => self.scan_template_escape_cr(),
-            CharacterCodes::LOWER_X => self.scan_template_hex_escape(),
-            CharacterCodes::LOWER_U => self.scan_template_unicode_escape(),
-            _ => Self::scan_template_unknown_escape(ch),
-        }
-    }
-
-    fn scan_template_escape_digit_zero(&mut self) -> String {
-        if self.pos < self.end && is_digit(self.char_code_unchecked(self.pos)) {
-            self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-            return String::from("\\0");
-        }
-        String::from("\0")
-    }
-
-    fn scan_template_escape_octal_digit(&mut self, ch: u32) -> String {
-        self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-        let digit = char::from_u32(ch).unwrap_or('?');
-        format!("\\{digit}")
-    }
-
-    fn scan_template_escape_cr(&mut self) -> String {
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::LINE_FEED {
-            self.pos += 1;
-        }
-        String::new()
-    }
-
-    fn scan_template_hex_escape(&mut self) -> String {
-        if self.pos + 2 <= self.end {
-            let hex = self.substring(self.pos, self.pos + 2);
-            if let Ok(code) = u32::from_str_radix(&hex, 16) {
-                self.pos += 2;
-                if let Some(c) = char::from_u32(code) {
-                    return c.to_string();
-                }
-            }
-        }
-        self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-        "\\x".to_string()
-    }
-
-    fn scan_template_unicode_escape(&mut self) -> String {
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::OPEN_BRACE {
-            return self.scan_template_brace_unicode_escape();
-        }
-
-        if self.pos + 4 <= self.end {
-            let hex = self.substring(self.pos, self.pos + 4);
-            if hex.len() == 4 && hex.bytes().all(|b| b.is_ascii_hexdigit()) {
-                let code = u32::from_str_radix(&hex, 16).unwrap_or(0);
-                self.pos += 4;
-                return Self::decode_fixed_unicode_code_unit(code);
-            }
-            self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-            return String::from("\\u");
-        }
-
-        self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-        String::from("\\u")
-    }
-
-    /// Decode the cooked value of a fixed four-hex-digit `\uXXXX` escape.
-    ///
-    /// A fixed four-hex-digit escape always denotes a single UTF-16 code unit in
-    /// the BMP range `0x0000..=0xFFFF`, so it is always a *valid* escape — even
-    /// when the value is a lone surrogate code unit (`0xD800..=0xDFFF`). `tsc`
-    /// treats such escapes as valid with a real cooked value, so they must not
-    /// set `ContainsInvalidEscape` (which would force tagged-template lowering).
-    ///
-    /// Lone surrogate code units have no standalone Unicode scalar value, so
-    /// they cannot be stored verbatim in a UTF-8 cooked string; they are cooked
-    /// to the replacement character. The cooked text only feeds the downlevel
-    /// cooked array, which this classification now correctly avoids emitting.
-    fn decode_fixed_unicode_code_unit(code: u32) -> String {
-        char::from_u32(code)
-            .unwrap_or(char::REPLACEMENT_CHARACTER)
-            .to_string()
-    }
-
-    fn scan_template_brace_unicode_escape(&mut self) -> String {
-        self.pos += 1;
-        let hex_start = self.pos;
-        while self.pos < self.end && is_hex_digit(self.char_code_unchecked(self.pos)) {
-            self.pos += 1;
-        }
-        if self.pos < self.end && self.char_code_unchecked(self.pos) == CharacterCodes::CLOSE_BRACE
-        {
-            let hex = self.substring(hex_start, self.pos);
-            self.pos += 1;
-            if let Ok(code) = u32::from_str_radix(&hex, 16)
-                && let Some(c) = char::from_u32(code)
-            {
-                return c.to_string();
-            }
-        }
-        self.token_flags |= TokenFlags::ContainsInvalidEscape as u32;
-        String::from("\\u")
-    }
-
-    fn scan_template_unknown_escape(ch: u32) -> String {
-        if let Some(c) = char::from_u32(ch) {
-            c.to_string()
-        } else {
-            String::new()
-        }
-    }
-
-    // =========================================================================
-    // JSX Scanning Methods
-    // =========================================================================
-
-    /// Scan a JSX identifier.
-    /// In JSX, identifiers can contain hyphens (like `data-testid`).
-    #[wasm_bindgen(js_name = scanJsxIdentifier)]
-    pub fn scan_jsx_identifier(&mut self) -> SyntaxKind {
-        if crate::token_is_identifier_or_keyword(self.token) {
-            let start = self.token_start;
-            let mut decoded = if (self.token_flags & TokenFlags::UnicodeEscape as u32) != 0 {
-                Some(self.get_token_value_ref().to_string())
-            } else {
-                None
-            };
-            let mut extended = false;
-            // Continue scanning to include any hyphenated parts.
-            // JSX identifiers can be like: foo-bar-baz, class-id, etc.
-            // Keywords like `class` can also start JSX attribute names.
-            while self.pos < self.end {
-                let ch = self.char_code_unchecked(self.pos);
-                if ch == CharacterCodes::MINUS {
-                    extended = true;
-                    // In JSX, hyphens are allowed in identifiers
-                    self.pos += 1;
-                    if let Some(text) = decoded.as_mut() {
-                        text.push('-');
-                    }
-                    // After a JSX hyphen, continuation may start with any identifier part.
-                    if self.pos >= self.end {
-                        continue;
-                    }
-                    let part_start = self.char_code_unchecked(self.pos);
-                    if part_start == CharacterCodes::BACKSLASH {
-                        if let Some(code_point) = self.peek_unicode_escape()
-                            && self.is_identifier_part(code_point)
-                        {
-                            let text =
-                                decoded.get_or_insert_with(|| self.source[start..self.pos].into());
-                            if let Some(c) =
-                                char::from_u32(self.scan_unicode_escape_value().unwrap_or(0))
-                            {
-                                text.push(c);
-                            }
-                            while self.pos < self.end {
-                                let ch = self.char_code_unchecked(self.pos);
-                                if ch == CharacterCodes::BACKSLASH {
-                                    if let Some(code_point) = self.peek_unicode_escape()
-                                        && self.is_identifier_part(code_point)
-                                    {
-                                        if let Some(c) = char::from_u32(
-                                            self.scan_unicode_escape_value().unwrap_or(0),
-                                        ) {
-                                            text.push(c);
-                                        }
-                                        continue;
-                                    }
-                                    break;
-                                }
-                                if !self.is_identifier_part(ch) {
-                                    break;
-                                }
-                                if let Some(c) = char::from_u32(ch) {
-                                    text.push(c);
-                                }
-                                self.pos += self.char_len_at(self.pos);
-                            }
-                        }
-                    } else if self.is_identifier_part(part_start) {
-                        loop {
-                            let ch = self.char_code_unchecked(self.pos);
-                            if !self.is_identifier_part(ch) {
-                                break;
-                            }
-                            if let Some(text) = decoded.as_mut()
-                                && let Some(c) = char::from_u32(ch)
-                            {
-                                text.push(c);
-                            }
-                            self.pos += self.char_len_at(self.pos); // Handle multi-byte UTF-8
-                            if self.pos >= self.end {
-                                break;
-                            }
-                        }
-                    }
-                } else {
-                    break;
-                }
-            }
-            if extended {
-                if let Some(text) = decoded {
-                    self.token_atom = self.interner.intern(&text);
-                    self.token_flags |= TokenFlags::UnicodeEscape as u32;
-                } else {
-                    // ZERO-ALLOCATION: Intern directly from source slice, clear token_value
-                    self.token_atom = self.interner.intern(&self.source[start..self.pos]);
-                }
-                self.token_value.clear();
-                // After extending with hyphens, the token becomes an Identifier
-                self.token = SyntaxKind::Identifier;
-            }
-        }
-        self.token
-    }
-
-    /// Re-scan the current token as a JSX token.
-    /// Used when the parser enters JSX context and needs to rescan.
-    /// Must reset to `full_start_pos` (before trivia), not `token_start` (after trivia),
-    /// so that JSX text nodes include leading whitespace/newlines.
-    /// Matches tsc: `pos = tokenStart = fullStartPos;`
-    #[wasm_bindgen(js_name = reScanJsxToken)]
-    pub fn re_scan_jsx_token(&mut self, allow_multiline_jsx_text: bool) -> SyntaxKind {
-        // Remove any scanner diagnostics emitted at positions >= full_start_pos.
-        // The previous scan (in normal JS mode) may have produced false diagnostics
-        // (e.g., TS1351 for `7x` in JSX text content). Since we're rescanning this
-        // range in JSX mode, those diagnostics are invalid.
-        let rescan_start = self.full_start_pos;
-        self.scanner_diagnostics.retain(|d| d.pos < rescan_start);
-        self.pos = self.full_start_pos;
-        self.scan_jsx_token(allow_multiline_jsx_text)
-    }
-
-    /// Scan a JSX token (text, open element, close element, etc.)
-    fn scan_jsx_token(&mut self, allow_multiline_jsx_text: bool) -> SyntaxKind {
-        self.full_start_pos = self.pos;
-        self.token_start = self.pos;
-        // Clear stale atom from any prior identifier scan so get_token_value_ref()
-        // returns the freshly scanned JSX text, not the old interned identifier.
-        self.token_atom = Atom::NONE;
-
-        if self.pos >= self.end {
-            self.token = SyntaxKind::EndOfFileToken;
-            return self.token;
-        }
-
-        // A merge-conflict marker (`<<<<<<<`, `|||||||`, `=======`, `>>>>>>>`)
-        // can appear inside JSX child content. Check before the LESS_THAN
-        // angle-bracket path: a 7-`<` run at line start is a conflict marker,
-        // not the start of a nested JSX tag. Mirrors the regular-mode handler
-        // which checks `is_conflict_marker_trivia()` first for `<`/`=`/`>`/`|`.
-        if self.is_conflict_marker_trivia() {
-            self.scan_conflict_marker_trivia();
-            if self.skip_trivia {
-                return self.scan_jsx_token(allow_multiline_jsx_text);
-            }
-            self.token = SyntaxKind::ConflictMarkerTrivia;
-            return self.token;
-        }
-
-        let ch = self.char_code_unchecked(self.pos);
-
-        // Check for JSX opening/closing angle brackets
-        if ch == CharacterCodes::LESS_THAN {
-            // Check for </
-            if self.char_code_at(self.pos + 1) == Some(CharacterCodes::SLASH) {
-                self.pos += 2;
-                self.token = SyntaxKind::LessThanSlashToken;
-                return self.token;
-            }
-            self.pos += 1;
-            self.token = SyntaxKind::LessThanToken;
-            return self.token;
-        }
-
-        if ch == CharacterCodes::OPEN_BRACE {
-            self.pos += 1;
-            self.token = SyntaxKind::OpenBraceToken;
-            return self.token;
-        }
-
-        // Scan JSX text
-        let mut text = String::new();
-        while self.pos < self.end {
-            let c = self.char_code_unchecked(self.pos);
-
-            // Stop on JSX special characters
-            if c == CharacterCodes::OPEN_BRACE || c == CharacterCodes::LESS_THAN {
-                break;
-            }
-
-            // TS1382: bare `>` in JSX text
-            if c == CharacterCodes::GREATER_THAN {
-                self.scanner_diagnostics.push(ScannerDiagnostic {
-                    pos: self.pos,
-                    length: 1,
-                    message: diagnostic_messages::UNEXPECTED_TOKEN_DID_YOU_MEAN_OR_GT,
-                    code: diagnostic_codes::UNEXPECTED_TOKEN_DID_YOU_MEAN_OR_GT,
-                    args: Vec::new(),
-                });
-            }
-
-            // TS1381: bare `}` in JSX text
-            if c == CharacterCodes::CLOSE_BRACE {
-                self.scanner_diagnostics.push(ScannerDiagnostic {
-                    pos: self.pos,
-                    length: 1,
-                    message: diagnostic_messages::UNEXPECTED_TOKEN_DID_YOU_MEAN_OR_RBRACE,
-                    code: diagnostic_codes::UNEXPECTED_TOKEN_DID_YOU_MEAN_OR_RBRACE,
-                    args: Vec::new(),
-                });
-            }
-
-            // Handle newlines in JSX text
-            if is_line_break(c) {
-                if !allow_multiline_jsx_text {
-                    break;
-                }
-                self.token_flags |= TokenFlags::PrecedingLineBreak as u32;
-            }
-
-            if let Some(char) = char::from_u32(c) {
-                text.push(char);
-            }
-            self.pos += self.char_len_at(self.pos); // Handle multi-byte UTF-8
-        }
-
-        if !text.is_empty() {
-            self.token_value = text;
-            self.token = SyntaxKind::JsxText;
-            return self.token;
-        }
-
-        self.token = SyntaxKind::Unknown;
-        self.token
-    }
-
-    /// Scan a JSX attribute value (string literal or expression).
-    #[wasm_bindgen(js_name = scanJsxAttributeValue)]
-    pub fn scan_jsx_attribute_value(&mut self) -> SyntaxKind {
-        self.full_start_pos = self.pos;
-        self.token_flags = 0;
-
-        // Skip whitespace
-        while self.pos < self.end && is_white_space_single_line(self.char_code_unchecked(self.pos))
-        {
-            self.pos += 1;
-        }
-
-        self.token_start = self.pos;
-
-        if self.pos >= self.end {
-            self.token = SyntaxKind::EndOfFileToken;
-            return self.token;
-        }
-
-        let ch = self.char_code_unchecked(self.pos);
-
-        // String literal
-        if ch == CharacterCodes::DOUBLE_QUOTE || ch == CharacterCodes::SINGLE_QUOTE {
-            self.scan_jsx_string_literal(ch);
-            return self.token;
-        }
-
-        self.scan()
-    }
-
-    /// Scan a JSX string literal (used for attribute values).
-    /// Unlike regular strings, JSX strings don't support escape sequences.
-    fn scan_jsx_string_literal(&mut self, quote: u32) {
-        self.pos += 1; // Skip opening quote
-        let mut result = String::new();
-
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-            if ch == quote {
-                self.pos += 1; // Closing quote
-                self.token_value = result;
-                self.token = SyntaxKind::StringLiteral;
-                return;
-            }
-            // JSX strings don't process escape sequences - they're literal
-            if let Some(c) = char::from_u32(ch) {
-                result.push(c);
-            }
-            // Advance by the character's UTF-8 byte length so multi-byte chars
-            // are not re-decoded from a continuation-byte offset.
-            self.pos += self.char_len_at(self.pos);
-        }
-
-        // Unterminated string
-        self.token_flags |= TokenFlags::Unterminated as u32;
-        self.token_value = result;
-        self.token = SyntaxKind::StringLiteral;
-    }
-
-    /// Re-scan a JSX attribute value from the current token position.
-    #[wasm_bindgen(js_name = reScanJsxAttributeValue)]
-    pub fn re_scan_jsx_attribute_value(&mut self) -> SyntaxKind {
-        self.pos = self.token_start;
-        self.scan_jsx_attribute_value()
-    }
-
-    // =========================================================================
-    // JSDoc Scanning Methods
-    // =========================================================================
-
-    /// Scan a `JSDoc` token.
-    /// Used when parsing `JSDoc` comments.
-    #[wasm_bindgen(js_name = scanJsDocToken)]
-    pub fn scan_jsdoc_token(&mut self) -> SyntaxKind {
-        self.full_start_pos = self.pos;
-        self.token_flags = 0;
-
-        if self.pos >= self.end {
-            self.token = SyntaxKind::EndOfFileToken;
-            return self.token;
-        }
-
-        self.token_start = self.pos;
-        let ch = self.char_code_unchecked(self.pos);
-
-        // Handle newlines
-        if ch == CharacterCodes::LINE_FEED || ch == CharacterCodes::CARRIAGE_RETURN {
-            self.token_flags |= TokenFlags::PrecedingLineBreak as u32;
-            self.pos += 1;
-            if ch == CharacterCodes::CARRIAGE_RETURN
-                && self.pos < self.end
-                && self.char_code_unchecked(self.pos) == CharacterCodes::LINE_FEED
-            {
-                self.pos += 1;
-            }
-            self.token = SyntaxKind::NewLineTrivia;
-            return self.token;
-        }
-
-        // Handle whitespace
-        if is_white_space_single_line(ch) {
-            while self.pos < self.end
-                && is_white_space_single_line(self.char_code_unchecked(self.pos))
-            {
-                self.pos += 1;
-            }
-            self.token = SyntaxKind::WhitespaceTrivia;
-            return self.token;
-        }
-
-        if self.scan_jsdoc_punctuation_token(ch) {
-            return self.token;
-        }
-
-        // Check for identifier
-        if self.is_identifier_start(ch) {
-            return self.scan_jsdoc_identifier();
-        }
-
-        // Unknown character - advance and return Unknown (properly handle multi-byte UTF-8)
-        self.scan_jsdoc_unknown_character();
-        self.token
-    }
-
-    fn scan_jsdoc_punctuation_token(&mut self, ch: u32) -> bool {
-        match ch {
-            CharacterCodes::AT => {
-                self.pos += 1;
-                self.token = SyntaxKind::AtToken;
-            }
-            CharacterCodes::ASTERISK => {
-                self.pos += 1;
-                self.token = SyntaxKind::AsteriskToken;
-            }
-            CharacterCodes::OPEN_BRACE => {
-                self.pos += 1;
-                self.token = SyntaxKind::OpenBraceToken;
-            }
-            CharacterCodes::CLOSE_BRACE => {
-                self.pos += 1;
-                self.token = SyntaxKind::CloseBraceToken;
-            }
-            CharacterCodes::OPEN_BRACKET => {
-                self.pos += 1;
-                self.token = SyntaxKind::OpenBracketToken;
-            }
-            CharacterCodes::CLOSE_BRACKET => {
-                self.pos += 1;
-                self.token = SyntaxKind::CloseBracketToken;
-            }
-            CharacterCodes::LESS_THAN => {
-                self.pos += 1;
-                self.token = SyntaxKind::LessThanToken;
-            }
-            CharacterCodes::GREATER_THAN => {
-                self.pos += 1;
-                self.token = SyntaxKind::GreaterThanToken;
-            }
-            CharacterCodes::EQUALS => {
-                self.pos += 1;
-                self.token = SyntaxKind::EqualsToken;
-            }
-            CharacterCodes::COMMA => {
-                self.pos += 1;
-                self.token = SyntaxKind::CommaToken;
-            }
-            CharacterCodes::DOT => {
-                self.pos += 1;
-                self.token = SyntaxKind::DotToken;
-            }
-            CharacterCodes::BACKTICK => {
-                self.pos += 1;
-                while self.pos < self.end
-                    && self.char_code_unchecked(self.pos) != CharacterCodes::BACKTICK
-                {
-                    self.pos += 1;
-                }
-                if self.pos < self.end {
-                    self.pos += 1;
-                }
-                self.token_value = self.substring(self.token_start, self.pos);
-                self.token = SyntaxKind::NoSubstitutionTemplateLiteral;
-            }
-            _ => return false,
-        }
-        true
-    }
-
-    fn scan_jsdoc_identifier(&mut self) -> SyntaxKind {
-        self.pos += self.char_len_at(self.pos);
-        while self.pos < self.end && self.is_identifier_part(self.char_code_unchecked(self.pos)) {
-            self.pos += self.char_len_at(self.pos);
-        }
-        self.token_value = self.substring(self.token_start, self.pos);
-        self.token = crate::text_to_keyword(&self.token_value).unwrap_or(SyntaxKind::Identifier);
-        self.token
-    }
-
-    fn scan_jsdoc_unknown_character(&mut self) {
-        self.pos += self.char_len_at(self.pos);
-        self.token = SyntaxKind::Unknown;
-    }
-
-    /// Scan `JSDoc` comment text token.
-    /// Used for scanning the text content within `JSDoc` comments.
-    #[wasm_bindgen(js_name = scanJsDocCommentTextToken)]
-    pub fn scan_jsdoc_comment_text_token(&mut self, in_backticks: bool) -> SyntaxKind {
-        self.full_start_pos = self.pos;
-        self.token_flags = 0;
-        self.token_start = self.pos;
-
-        if self.pos >= self.end {
-            self.token = SyntaxKind::EndOfFileToken;
-            return self.token;
-        }
-
-        // Scan until we hit a special character
-        while self.pos < self.end {
-            let ch = self.char_code_unchecked(self.pos);
-
-            // Check for end markers
-            match ch {
-                // Always stop on newline
-                CharacterCodes::LINE_FEED | CharacterCodes::CARRIAGE_RETURN => {
-                    break;
-                }
-                // Stop on @ unless in backticks
-                CharacterCodes::AT | CharacterCodes::OPEN_BRACE | CharacterCodes::CLOSE_BRACE
-                    if !in_backticks =>
-                {
-                    break;
-                }
-                // Stop on backtick - it toggles the mode
-                CharacterCodes::BACKTICK => {
-                    if self.pos > self.token_start {
-                        break; // Return text first
-                    }
-                    // Return the backtick token
-                    self.pos += 1;
-                    self.token = SyntaxKind::Unknown; // Use Unknown to signal backtick
-                    return self.token;
-                }
-                _ => {
-                    // Properly handle multi-byte UTF-8 characters
-                    self.pos += self.char_len_at(self.pos);
-                }
-            }
-        }
-
-        if self.pos > self.token_start {
-            self.token_value = self.substring(self.token_start, self.pos);
-            // JSDocText token would be returned here, but we use Identifier as a stand-in
-            self.token = SyntaxKind::Identifier;
-        } else {
-            self.token = SyntaxKind::EndOfFileToken;
-        }
-        self.token
-    }
-
-    // =========================================================================
-    // Shebang Handling
-    // =========================================================================
-
-    /// Scan a shebang (#!) at the start of the file.
-    /// Returns the length of the shebang line (including newline), or 0 if no shebang.
-    #[wasm_bindgen(js_name = scanShebangTrivia)]
-    pub fn scan_shebang_trivia(&mut self) -> usize {
-        // Shebang must be at the very start of the file
-        if self.pos != 0 {
-            return 0;
-        }
-
-        // Check for #!
-        if self.pos + 1 < self.end
-            && self.char_code_unchecked(self.pos) == CharacterCodes::HASH
-            && self.char_code_unchecked(self.pos + 1) == CharacterCodes::EXCLAMATION
-        {
-            let start = self.pos;
-            self.pos += 2;
-
-            // Scan to end of line
-            while self.pos < self.end {
-                let ch = self.char_code_unchecked(self.pos);
-                if ch == CharacterCodes::LINE_FEED || ch == CharacterCodes::CARRIAGE_RETURN {
-                    break;
-                }
-                // Use char_len_at to properly handle multi-byte UTF-8 characters
-                self.pos += self.char_len_at(self.pos);
-            }
-
-            // Include the newline in the shebang
-            if self.pos < self.end {
-                let ch = self.char_code_unchecked(self.pos);
-                if ch == CharacterCodes::CARRIAGE_RETURN {
-                    self.pos += 1;
-                    if self.pos < self.end
-                        && self.char_code_unchecked(self.pos) == CharacterCodes::LINE_FEED
-                    {
-                        self.pos += 1;
-                    }
-                } else if ch == CharacterCodes::LINE_FEED {
-                    self.pos += 1;
-                }
-            }
-
-            return self.pos - start;
-        }
-
-        0
-    }
 }
 
 // =============================================================================
@@ -3449,742 +1460,4 @@ const fn is_regex_flag(ch: u32) -> bool {
 // =============================================================================
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Helper: create a scanner that skips trivia and collect all tokens.
-    fn scan_all(source: &str) -> Vec<(SyntaxKind, String)> {
-        let mut scanner = ScannerState::new(source.to_string(), true);
-        let mut tokens = Vec::new();
-        loop {
-            let kind = scanner.scan();
-            if kind == SyntaxKind::EndOfFileToken {
-                break;
-            }
-            tokens.push((kind, scanner.get_token_value()));
-        }
-        tokens
-    }
-
-    /// Helper: create a scanner that preserves trivia and collect all tokens.
-    fn scan_all_with_trivia(source: &str) -> Vec<(SyntaxKind, String)> {
-        let mut scanner = ScannerState::new(source.to_string(), false);
-        let mut tokens = Vec::new();
-        loop {
-            let kind = scanner.scan();
-            if kind == SyntaxKind::EndOfFileToken {
-                break;
-            }
-            tokens.push((kind, scanner.get_token_text()));
-        }
-        tokens
-    }
-
-    // ── Empty input ───────────────────────────────────────────────────
-
-    #[test]
-    fn empty_input_returns_eof() {
-        let mut scanner = ScannerState::new(String::new(), true);
-        assert_eq!(scanner.scan(), SyntaxKind::EndOfFileToken);
-    }
-
-    // ── Identifiers ───────────────────────────────────────────────────
-
-    #[test]
-    fn scan_identifiers() {
-        let tokens = scan_all("foo bar _baz $qux");
-        assert_eq!(tokens.len(), 4);
-        assert_eq!(tokens[0], (SyntaxKind::Identifier, "foo".to_string()));
-        assert_eq!(tokens[1], (SyntaxKind::Identifier, "bar".to_string()));
-        assert_eq!(tokens[2], (SyntaxKind::Identifier, "_baz".to_string()));
-        assert_eq!(tokens[3], (SyntaxKind::Identifier, "$qux".to_string()));
-    }
-
-    #[test]
-    fn scan_identifier_with_other_id_continue_middle_dot() {
-        let tokens = scan_all("a·b");
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(tokens[0], (SyntaxKind::Identifier, "a·b".to_string()));
-    }
-
-    #[test]
-    fn scan_es2015_braced_astral_escape_as_identifier_start() {
-        let mut scanner = ScannerState::new(r"\u{102A7}tail".to_string(), true);
-        scanner.set_language_version(ScriptTarget::ES2015);
-
-        assert_eq!(scanner.scan(), SyntaxKind::Identifier);
-        assert_eq!(scanner.get_token_value_ref(), "𐊧tail");
-        assert_eq!(scanner.get_token_text(), r"\u{102A7}tail");
-    }
-
-    #[test]
-    fn scan_es5_braced_astral_escape_remains_invalid_identifier_start() {
-        let mut scanner = ScannerState::new(r"\u{102A7}tail".to_string(), true);
-        scanner.set_language_version(ScriptTarget::ES5);
-
-        assert_eq!(scanner.scan(), SyntaxKind::Unknown);
-        assert_eq!(scanner.get_token_text(), "\\");
-    }
-
-    // ── Keywords ──────────────────────────────────────────────────────
-
-    #[test]
-    fn scan_keywords() {
-        let tokens = scan_all("if else while for const let var");
-        assert_eq!(tokens.len(), 7);
-        assert_eq!(tokens[0].0, SyntaxKind::IfKeyword);
-        assert_eq!(tokens[1].0, SyntaxKind::ElseKeyword);
-        assert_eq!(tokens[2].0, SyntaxKind::WhileKeyword);
-        assert_eq!(tokens[3].0, SyntaxKind::ForKeyword);
-        assert_eq!(tokens[4].0, SyntaxKind::ConstKeyword);
-        assert_eq!(tokens[5].0, SyntaxKind::LetKeyword);
-        assert_eq!(tokens[6].0, SyntaxKind::VarKeyword);
-    }
-
-    // ── Numeric literals ──────────────────────────────────────────────
-
-    #[test]
-    fn scan_numeric_literals() {
-        let tokens = scan_all("0 42 3.14 0xFF 0b1010 0o777 1_000");
-        assert_eq!(tokens.len(), 7);
-        for (kind, _) in &tokens {
-            assert_eq!(*kind, SyntaxKind::NumericLiteral);
-        }
-        assert_eq!(tokens[0].1, "0");
-        assert_eq!(tokens[1].1, "42");
-        assert_eq!(tokens[2].1, "3.14");
-        assert_eq!(tokens[3].1, "0xFF");
-        assert_eq!(tokens[4].1, "0b1010");
-        assert_eq!(tokens[5].1, "0o777");
-        assert_eq!(tokens[6].1, "1_000");
-    }
-
-    #[test]
-    fn scan_bigint_suffix_on_failed_exponent_preserves_full_text() {
-        // Regression for `identifierStartAfterNumericLiteral`: when a numeric
-        // literal has a failed exponent immediately followed by `n` (e.g.
-        // `3en`, `123en`), tsc treats the `n` as a recovered BigInt suffix
-        // and the literal text spans the full `<digits>en` range. Emit must
-        // preserve that source spelling — printing `3e` instead drops the
-        // `n` and produces `3e[null];` where tsc produces `3en[null];`.
-        let tokens = scan_all("3en 123en");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].0, SyntaxKind::NumericLiteral);
-        assert_eq!(tokens[0].1, "3en");
-        assert_eq!(tokens[1].0, SyntaxKind::NumericLiteral);
-        assert_eq!(tokens[1].1, "123en");
-
-        // Sanity: `1ee` is a different shape — the trailing `e` is not an
-        // `n` BigInt suffix, so the scanner resets pos before `e` and the
-        // numeric token is just `1e`. Emit then renders `1e; e;`.
-        let tokens = scan_all("1ee");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].0, SyntaxKind::NumericLiteral);
-        assert_eq!(tokens[0].1, "1e");
-        assert_eq!(tokens[1].0, SyntaxKind::Identifier);
-        assert_eq!(tokens[1].1, "e");
-    }
-
-    // ── String literals ───────────────────────────────────────────────
-
-    #[test]
-    fn scan_string_literals() {
-        let tokens = scan_all(r#""hello" 'world'"#);
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].0, SyntaxKind::StringLiteral);
-        assert_eq!(tokens[0].1, "hello");
-        assert_eq!(tokens[1].0, SyntaxKind::StringLiteral);
-        assert_eq!(tokens[1].1, "world");
-    }
-
-    #[test]
-    fn scan_invalid_hex_escape_emits_ts1125() {
-        // Regression for `compiler/stringLiteralsErrors.ts`: tsc emits
-        // TS1125 "Hexadecimal digit expected." for `\x` followed by fewer
-        // than two hex digits. Anchor is at the position the scan halted
-        // (the first non-hex char or the closing quote).
-        let mut scanner = ScannerState::new(r#""\x0""#.to_string(), true);
-        loop {
-            if scanner.scan() == SyntaxKind::EndOfFileToken {
-                break;
-            }
-        }
-        let diags = &scanner.scanner_diagnostics;
-        // `"` is at byte 0, `\` at 1, `x` at 2, `0` at 3, closing `"` at 4.
-        // After consuming the single hex digit `0`, scanner halts at the
-        // closing quote (byte 4). tsc emits the error there.
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.code == diagnostic_codes::HEXADECIMAL_DIGIT_EXPECTED && d.pos == 4),
-            "expected TS1125 at the closing quote (byte 4), got: {diags:?}"
-        );
-
-        let mut scanner2 = ScannerState::new(r#""\xmm""#.to_string(), true);
-        loop {
-            if scanner2.scan() == SyntaxKind::EndOfFileToken {
-                break;
-            }
-        }
-        let diags2 = &scanner2.scanner_diagnostics;
-        // `\xmm`: scanner halts at the first `m` (byte 3) since it's not hex.
-        assert!(
-            diags2
-                .iter()
-                .any(|d| d.code == diagnostic_codes::HEXADECIMAL_DIGIT_EXPECTED && d.pos == 3),
-            "expected TS1125 at the first non-hex char (byte 3), got: {diags2:?}"
-        );
-
-        // Sanity guard: a well-formed `\x41` ('A') must not gain a diagnostic.
-        let mut scanner3 = ScannerState::new(r#""\x41""#.to_string(), true);
-        loop {
-            if scanner3.scan() == SyntaxKind::EndOfFileToken {
-                break;
-            }
-        }
-        assert!(
-            !scanner3
-                .scanner_diagnostics
-                .iter()
-                .any(|d| d.code == diagnostic_codes::HEXADECIMAL_DIGIT_EXPECTED),
-            "well-formed `\\x41` must not emit TS1125, got: {:?}",
-            scanner3.scanner_diagnostics
-        );
-    }
-
-    // ── Template literals ─────────────────────────────────────────────
-
-    #[test]
-    fn scan_no_substitution_template() {
-        let tokens = scan_all("`hello world`");
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(tokens[0].0, SyntaxKind::NoSubstitutionTemplateLiteral);
-        assert_eq!(tokens[0].1, "hello world");
-    }
-
-    // ── JSX attribute string literals ─────────────────────────────────
-
-    /// Regression for #3977: `scan_jsx_string_literal` advanced by one byte
-    /// after pushing a character, so multi-byte UTF-8 chars were re-decoded
-    /// from a continuation-byte offset and pushed again. The token value of
-    /// `"é"` became `"éé"`, breaking JSX attribute string-literal types.
-    #[test]
-    fn scan_jsx_attribute_value_preserves_non_ascii() {
-        // 2-byte UTF-8 (é = U+00E9, 0xC3 0xA9).
-        let mut s = ScannerState::new("\"é\"".to_string(), true);
-        assert_eq!(s.scan_jsx_attribute_value(), SyntaxKind::StringLiteral);
-        assert_eq!(s.get_token_value(), "é");
-
-        // 3-byte UTF-8 (中 = U+4E2D).
-        let mut s = ScannerState::new("\"中文\"".to_string(), true);
-        assert_eq!(s.scan_jsx_attribute_value(), SyntaxKind::StringLiteral);
-        assert_eq!(s.get_token_value(), "中文");
-
-        // 4-byte UTF-8 (😀 = U+1F600, surrogate-pair scalar).
-        let mut s = ScannerState::new("\"a😀b\"".to_string(), true);
-        assert_eq!(s.scan_jsx_attribute_value(), SyntaxKind::StringLiteral);
-        assert_eq!(s.get_token_value(), "a😀b");
-
-        // Mixed widths and single quotes also exercised by JSX attributes.
-        let mut s = ScannerState::new("'héllo 中 😀'".to_string(), true);
-        assert_eq!(s.scan_jsx_attribute_value(), SyntaxKind::StringLiteral);
-        assert_eq!(s.get_token_value(), "héllo 中 😀");
-    }
-
-    // ── Punctuation ───────────────────────────────────────────────────
-
-    #[test]
-    fn scan_punctuation() {
-        let tokens = scan_all("{ } ( ) [ ] ; , . ...");
-        let kinds: Vec<SyntaxKind> = tokens.iter().map(|(k, _)| *k).collect();
-        assert_eq!(
-            kinds,
-            vec![
-                SyntaxKind::OpenBraceToken,
-                SyntaxKind::CloseBraceToken,
-                SyntaxKind::OpenParenToken,
-                SyntaxKind::CloseParenToken,
-                SyntaxKind::OpenBracketToken,
-                SyntaxKind::CloseBracketToken,
-                SyntaxKind::SemicolonToken,
-                SyntaxKind::CommaToken,
-                SyntaxKind::DotToken,
-                SyntaxKind::DotDotDotToken,
-            ]
-        );
-    }
-
-    // ── Operators ─────────────────────────────────────────────────────
-
-    #[test]
-    fn scan_comparison_operators() {
-        // Note: > is always scanned as GreaterThanToken; the parser calls
-        // re_scan_greater_token() to disambiguate >= / >> / >>> etc.
-        let tokens = scan_all("== != === !== < <=");
-        let kinds: Vec<SyntaxKind> = tokens.iter().map(|(k, _)| *k).collect();
-        assert_eq!(
-            kinds,
-            vec![
-                SyntaxKind::EqualsEqualsToken,
-                SyntaxKind::ExclamationEqualsToken,
-                SyntaxKind::EqualsEqualsEqualsToken,
-                SyntaxKind::ExclamationEqualsEqualsToken,
-                SyntaxKind::LessThanToken,
-                SyntaxKind::LessThanEqualsToken,
-            ]
-        );
-    }
-
-    #[test]
-    fn scan_greater_than_is_single_token() {
-        // The scanner always produces GreaterThanToken for '>'.
-        // Multi-char variants (>=, >>, >>>, >>=, >>>=) require re_scan_greater_token().
-        let tokens = scan_all(">");
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(tokens[0].0, SyntaxKind::GreaterThanToken);
-
-        // ">=" is scanned as ">" then "=" by the raw scanner
-        let tokens = scan_all(">=");
-        assert_eq!(tokens[0].0, SyntaxKind::GreaterThanToken);
-    }
-
-    #[test]
-    fn scan_assignment_operators() {
-        // Exclude >>= and >>>= which need re_scan_greater_token from parser context
-        let tokens = scan_all("= += -= *= **= /= %= <<= &= |= ^= ||= &&= ??=");
-        let kinds: Vec<SyntaxKind> = tokens.iter().map(|(k, _)| *k).collect();
-        assert_eq!(
-            kinds,
-            vec![
-                SyntaxKind::EqualsToken,
-                SyntaxKind::PlusEqualsToken,
-                SyntaxKind::MinusEqualsToken,
-                SyntaxKind::AsteriskEqualsToken,
-                SyntaxKind::AsteriskAsteriskEqualsToken,
-                SyntaxKind::SlashEqualsToken,
-                SyntaxKind::PercentEqualsToken,
-                SyntaxKind::LessThanLessThanEqualsToken,
-                SyntaxKind::AmpersandEqualsToken,
-                SyntaxKind::BarEqualsToken,
-                SyntaxKind::CaretEqualsToken,
-                SyntaxKind::BarBarEqualsToken,
-                SyntaxKind::AmpersandAmpersandEqualsToken,
-                SyntaxKind::QuestionQuestionEqualsToken,
-            ]
-        );
-    }
-
-    #[test]
-    fn scan_logical_operators() {
-        let tokens = scan_all("&& || ?? !");
-        let kinds: Vec<SyntaxKind> = tokens.iter().map(|(k, _)| *k).collect();
-        assert_eq!(
-            kinds,
-            vec![
-                SyntaxKind::AmpersandAmpersandToken,
-                SyntaxKind::BarBarToken,
-                SyntaxKind::QuestionQuestionToken,
-                SyntaxKind::ExclamationToken,
-            ]
-        );
-    }
-
-    #[test]
-    fn scan_arrow_function() {
-        let tokens = scan_all("=>");
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(tokens[0].0, SyntaxKind::EqualsGreaterThanToken);
-    }
-
-    // ── Trivia handling ───────────────────────────────────────────────
-
-    #[test]
-    fn trivia_skip_mode() {
-        let tokens = scan_all("  a  b  ");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].1, "a");
-        assert_eq!(tokens[1].1, "b");
-    }
-
-    #[test]
-    fn trivia_preserve_mode() {
-        let tokens = scan_all_with_trivia(" a ");
-        assert_eq!(tokens.len(), 3);
-        assert_eq!(tokens[0].0, SyntaxKind::WhitespaceTrivia);
-        assert_eq!(tokens[1].0, SyntaxKind::Identifier);
-        assert_eq!(tokens[2].0, SyntaxKind::WhitespaceTrivia);
-    }
-
-    #[test]
-    fn newline_trivia_preserved() {
-        let tokens = scan_all_with_trivia("a\nb");
-        assert_eq!(tokens.len(), 3);
-        assert_eq!(tokens[0].0, SyntaxKind::Identifier);
-        assert_eq!(tokens[1].0, SyntaxKind::NewLineTrivia);
-        assert_eq!(tokens[2].0, SyntaxKind::Identifier);
-    }
-
-    // ── Comments ──────────────────────────────────────────────────────
-
-    #[test]
-    fn single_line_comment_skipped() {
-        let tokens = scan_all("a // comment\nb");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].1, "a");
-        assert_eq!(tokens[1].1, "b");
-    }
-
-    #[test]
-    fn multi_line_comment_skipped() {
-        let tokens = scan_all("a /* comment */ b");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].1, "a");
-        assert_eq!(tokens[1].1, "b");
-    }
-
-    #[test]
-    fn single_line_comment_terminates_at_unicode_line_separator() {
-        let tokens = scan_all("a // comment\u{2028}b");
-        assert_eq!(tokens.len(), 2, "tokens were: {tokens:?}");
-        assert_eq!(tokens[0].1, "a");
-        assert_eq!(tokens[1].1, "b");
-    }
-
-    #[test]
-    fn single_line_comment_terminates_at_unicode_paragraph_separator() {
-        let tokens = scan_all("a // comment\u{2029}b");
-        assert_eq!(tokens.len(), 2, "tokens were: {tokens:?}");
-        assert_eq!(tokens[0].1, "a");
-        assert_eq!(tokens[1].1, "b");
-    }
-
-    #[test]
-    fn directive_comment_does_not_swallow_next_line_via_unicode_line_separator() {
-        let mut scanner =
-            ScannerState::new("// @ts-expect-error\u{2028}const ok = 1;".to_string(), true);
-        assert_eq!(scanner.scan(), SyntaxKind::ConstKeyword);
-        assert!(scanner.has_preceding_line_break());
-    }
-
-    #[test]
-    fn multi_line_comment_unicode_line_separator_sets_preceding_line_break() {
-        let mut scanner = ScannerState::new("a /*\u{2028}*/ b".to_string(), true);
-        scanner.scan(); // "a"
-        assert!(!scanner.has_preceding_line_break());
-        scanner.scan(); // "b"
-        assert!(scanner.has_preceding_line_break());
-    }
-
-    // ── Scanner state methods ─────────────────────────────────────────
-
-    #[test]
-    fn scanner_position_tracking() {
-        let mut scanner = ScannerState::new("abc def".to_string(), true);
-        scanner.scan(); // "abc"
-        assert_eq!(scanner.get_token_start(), 0);
-        assert_eq!(scanner.get_token_end(), 3);
-
-        scanner.scan(); // "def"
-        assert_eq!(scanner.get_token_start(), 4);
-        assert_eq!(scanner.get_token_end(), 7);
-    }
-
-    #[test]
-    fn scanner_set_text() {
-        let mut scanner = ScannerState::new("abc".to_string(), true);
-        scanner.scan();
-        assert_eq!(scanner.get_token_value(), "abc");
-
-        scanner.set_text("xyz".to_string(), None, None);
-        scanner.scan();
-        assert_eq!(scanner.get_token_value(), "xyz");
-    }
-
-    #[test]
-    fn scanner_reset_token_state() {
-        let mut scanner = ScannerState::new("ab cd".to_string(), true);
-        scanner.scan(); // "ab"
-        scanner.scan(); // "cd"
-        scanner.reset_token_state(0);
-        scanner.scan();
-        assert_eq!(scanner.get_token_value(), "ab");
-    }
-
-    // ── Preceding line break ──────────────────────────────────────────
-
-    #[test]
-    fn preceding_line_break_detection() {
-        let mut scanner = ScannerState::new("a\nb".to_string(), true);
-        scanner.scan(); // "a"
-        assert!(!scanner.has_preceding_line_break());
-        scanner.scan(); // "b"
-        assert!(scanner.has_preceding_line_break());
-    }
-
-    // ── BigInt literals ───────────────────────────────────────────────
-
-    #[test]
-    fn scan_bigint_literal() {
-        let tokens = scan_all("42n 0xFFn");
-        assert_eq!(tokens.len(), 2);
-        assert_eq!(tokens[0].0, SyntaxKind::BigIntLiteral);
-        assert_eq!(tokens[1].0, SyntaxKind::BigIntLiteral);
-    }
-
-    // ── Optional chaining ─────────────────────────────────────────────
-
-    #[test]
-    fn scan_optional_chaining() {
-        let tokens = scan_all("a?.b");
-        let kinds: Vec<SyntaxKind> = tokens.iter().map(|(k, _)| *k).collect();
-        assert_eq!(
-            kinds,
-            vec![
-                SyntaxKind::Identifier,
-                SyntaxKind::QuestionDotToken,
-                SyntaxKind::Identifier,
-            ]
-        );
-    }
-
-    // ── Hash/private identifier ───────────────────────────────────────
-
-    #[test]
-    fn scan_private_identifier() {
-        let tokens = scan_all("#field");
-        assert_eq!(tokens.len(), 1);
-        assert_eq!(tokens[0].0, SyntaxKind::PrivateIdentifier);
-        assert_eq!(tokens[0].1, "#field");
-    }
-
-    // ── Helper function tests ─────────────────────────────────────────
-
-    #[test]
-    fn is_line_break_chars() {
-        assert!(is_line_break(CharacterCodes::LINE_FEED));
-        assert!(is_line_break(CharacterCodes::CARRIAGE_RETURN));
-        assert!(is_line_break(CharacterCodes::LINE_SEPARATOR));
-        assert!(is_line_break(CharacterCodes::PARAGRAPH_SEPARATOR));
-        assert!(!is_line_break(CharacterCodes::SPACE));
-        assert!(!is_line_break(CharacterCodes::TAB));
-    }
-
-    #[test]
-    fn is_white_space_single_line_chars() {
-        assert!(is_white_space_single_line(CharacterCodes::SPACE));
-        assert!(is_white_space_single_line(CharacterCodes::TAB));
-        assert!(is_white_space_single_line(CharacterCodes::FORM_FEED));
-        assert!(is_white_space_single_line(
-            CharacterCodes::NON_BREAKING_SPACE
-        ));
-        assert!(!is_white_space_single_line(CharacterCodes::LINE_FEED));
-        assert!(!is_white_space_single_line(CharacterCodes::CARRIAGE_RETURN));
-        assert!(!is_white_space_single_line(0x41)); // 'A'
-    }
-
-    #[test]
-    fn is_identifier_start_chars() {
-        assert!(is_identifier_start(CharacterCodes::LOWER_A));
-        assert!(is_identifier_start(CharacterCodes::UPPER_Z));
-        assert!(is_identifier_start(CharacterCodes::UNDERSCORE));
-        assert!(is_identifier_start(CharacterCodes::DOLLAR));
-        assert!(!is_identifier_start(CharacterCodes::_0));
-        assert!(!is_identifier_start(CharacterCodes::SPACE));
-        assert!(!is_identifier_start(CharacterCodes::PLUS));
-    }
-
-    #[test]
-    fn is_identifier_part_rejects_subscript_digits() {
-        // U+2081 SUBSCRIPT ONE is No (Number, other), NOT Nd — must be rejected
-        assert!(!is_identifier_part(0x2081)); // ₁
-        assert!(!is_identifier_part(0x2082)); // ₂
-        assert!(!is_identifier_part(0x00B2)); // ² SUPERSCRIPT TWO (No)
-        assert!(!is_identifier_part(0x00B3)); // ³ SUPERSCRIPT THREE (No)
-        assert!(!is_identifier_part(0x00BC)); // ¼ VULGAR FRACTION ONE QUARTER (No)
-        // Nd digits should be accepted
-        assert!(is_identifier_part(0x0966)); // Devanagari digit zero (Nd)
-        assert!(is_identifier_part(0x0660)); // Arabic-Indic digit zero (Nd)
-        assert!(is_identifier_part(0xFF10)); // Fullwidth digit zero (Nd)
-        // ASCII digits should be accepted
-        assert!(is_identifier_part(0x30)); // '0'
-        assert!(is_identifier_part(0x39)); // '9'
-        // Letters should be accepted
-        assert!(is_identifier_part(0x61)); // 'a'
-        // Other_ID_Continue characters should be accepted
-        assert!(is_identifier_part(0x00B7)); // · MIDDLE DOT
-        assert!(is_identifier_part(0x0387)); // · GREEK ANO TELEIA
-        assert!(is_identifier_part(0x1369)); // ፩ ETHIOPIC DIGIT ONE
-        assert!(is_identifier_part(0x19DA)); // ᧚ NEW TAI LUE THAM DIGIT ONE
-    }
-
-    #[test]
-    fn is_digit_chars() {
-        assert!(is_digit(CharacterCodes::_0));
-        assert!(is_digit(CharacterCodes::_9));
-        assert!(!is_digit(CharacterCodes::LOWER_A));
-        assert!(!is_digit(CharacterCodes::SPACE));
-    }
-
-    #[test]
-    fn is_regex_flag_chars() {
-        assert!(is_regex_flag(CharacterCodes::LOWER_G));
-        assert!(is_regex_flag(CharacterCodes::LOWER_I));
-        assert!(is_regex_flag(CharacterCodes::LOWER_M));
-        assert!(is_regex_flag(CharacterCodes::LOWER_S));
-        assert!(is_regex_flag(CharacterCodes::LOWER_U));
-        assert!(is_regex_flag(CharacterCodes::LOWER_V));
-        assert!(is_regex_flag(CharacterCodes::LOWER_Y));
-        assert!(is_regex_flag(CharacterCodes::LOWER_D));
-        assert!(!is_regex_flag(CharacterCodes::LOWER_A));
-        assert!(!is_regex_flag(CharacterCodes::LOWER_Z));
-    }
-
-    // ── Scanner snapshot/restore ──────────────────────────────────────
-
-    #[test]
-    fn scanner_snapshot_and_restore() {
-        let mut scanner = ScannerState::new("a + b".to_string(), true);
-        scanner.scan(); // "a"
-        let snapshot = scanner.save_state();
-        scanner.scan(); // "+"
-        scanner.scan(); // "b"
-        assert_eq!(scanner.get_token_value(), "b");
-        scanner.restore_state(snapshot);
-        assert_eq!(scanner.get_token(), SyntaxKind::Identifier);
-        // After restoring, scanning again should give "+"
-        let next = scanner.scan();
-        assert_eq!(next, SyntaxKind::PlusToken);
-    }
-
-    // ── Rescan methods ────────────────────────────────────────────────
-
-    #[test]
-    fn rescan_greater_than_token() {
-        // The scanner always scans ">" as GreaterThanToken.
-        // re_scan_greater_token() is used by the parser to check if it could be >=, >>, etc.
-        let mut scanner = ScannerState::new(">= x".to_string(), true);
-        scanner.scan(); // scans ">"
-        assert_eq!(scanner.get_token(), SyntaxKind::GreaterThanToken);
-
-        // Rescan to get >=
-        let rescanned = scanner.re_scan_greater_token();
-        assert_eq!(rescanned, SyntaxKind::GreaterThanEqualsToken);
-    }
-
-    // ── Numeric separator diagnostics ────────────────────────────────
-    //
-    // These lock in tsc parity for `scanHexDigits`/`scanNumberFragment`:
-    // each invalid `_` produces its own diagnostic, distinguished by whether
-    // the preceding character was a *valid* separator (TS6189) or not
-    // (TS6188). Same-position emissions collapse to the first, mirroring
-    // tsc's `parseErrorAtPosition` skip-if-same-start dedup.
-
-    fn scan_separator_diagnostics(source: &str) -> Vec<(usize, u32)> {
-        let mut scanner = ScannerState::new(source.to_string(), true);
-        loop {
-            let kind = scanner.scan();
-            if kind == SyntaxKind::EndOfFileToken {
-                break;
-            }
-        }
-        scanner
-            .get_scanner_diagnostics()
-            .iter()
-            .map(|d| (d.pos, d.code))
-            .collect()
-    }
-
-    #[test]
-    fn separator_three_leading_underscores_emits_three_ts6188() {
-        // `0x___0111010_0101_1` — tsc reports TS6188 at each of the three
-        // leading underscores. Previously we only recorded the first one.
-        let diags = scan_separator_diagnostics("0x___0111010_0101_1");
-        let ts6188 = diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE;
-        assert_eq!(
-            diags,
-            vec![(2, ts6188), (3, ts6188), (4, ts6188)],
-            "expected three TS6188 at positions 2,3,4 for `0x___...`",
-        );
-    }
-
-    #[test]
-    fn separator_trailing_double_underscore_emits_one_ts6189() {
-        // `0X0110_0110__` — tsc emits TS6189 at the second trailing `_`
-        // and tsc's parser dedups the would-be TS6188 trailing-underscore
-        // companion at the same byte. Mirror the dedup at the scanner.
-        let diags = scan_separator_diagnostics("0X0110_0110__");
-        let ts6189 = diagnostic_codes::MULTIPLE_CONSECUTIVE_NUMERIC_SEPARATORS_ARE_NOT_PERMITTED;
-        assert_eq!(diags, vec![(12, ts6189)]);
-    }
-
-    #[test]
-    fn separator_consecutive_after_valid_emits_ts6189() {
-        // `0x01__11` — first `_` valid, second `_` is consecutive → TS6189.
-        let diags = scan_separator_diagnostics("0x01__11");
-        let ts6189 = diagnostic_codes::MULTIPLE_CONSECUTIVE_NUMERIC_SEPARATORS_ARE_NOT_PERMITTED;
-        assert_eq!(diags, vec![(5, ts6189)]);
-    }
-
-    #[test]
-    fn separator_leading_emits_ts6188() {
-        // `0x_110` — leading `_` after the prefix is invalid (TS6188).
-        let diags = scan_separator_diagnostics("0x_110");
-        let ts6188 = diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE;
-        assert_eq!(diags, vec![(2, ts6188)]);
-    }
-
-    #[test]
-    fn separator_trailing_single_underscore_emits_ts6188() {
-        // `0x00_` — trailing `_` only fires the post-loop trailing check.
-        let diags = scan_separator_diagnostics("0x00_");
-        let ts6188 = diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE;
-        assert_eq!(diags, vec![(4, ts6188)]);
-    }
-
-    #[test]
-    fn separator_valid_does_not_emit() {
-        // `1_000_000` and `0xFF_FF` are well-formed.
-        assert!(scan_separator_diagnostics("1_000_000").is_empty());
-        assert!(scan_separator_diagnostics("0xFF_FF").is_empty());
-    }
-
-    #[test]
-    fn separator_after_leading_zero_emits_ts6188() {
-        // `0_0` — separator immediately after a leading zero in an unprefixed
-        // numeric literal is a legacy-octal-style pattern that tsc rejects
-        // with TS6188 at the `_` position. The literal still tokenizes as a
-        // NumericLiteral so the rest of the parse can proceed; only the
-        // separator-not-allowed-here diagnostic is added.
-        let ts6188 = diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE;
-        assert_eq!(scan_separator_diagnostics("0_0"), vec![(1, ts6188)]);
-        assert_eq!(scan_separator_diagnostics("0_1"), vec![(1, ts6188)]);
-        assert_eq!(scan_separator_diagnostics("0_8"), vec![(1, ts6188)]);
-        // Also fires when followed by a fraction or exponent.
-        assert_eq!(scan_separator_diagnostics("0_0.5_5"), vec![(1, ts6188)]);
-        assert_eq!(scan_separator_diagnostics("0_0e5_5"), vec![(1, ts6188)]);
-    }
-
-    #[test]
-    fn separator_after_leading_zero_followed_by_double_underscore() {
-        // Regression: `parser.numericSeparators.decmialNegative.ts` files 18,
-        // 31, 44 (`0__0.0e0`, `0__0.0e+0`, `0__0.0e-0`). The leading-zero
-        // TS6188 rule must fire when the `_` after `0` is followed by another
-        // `_` (which itself starts a consecutive-separator run that ends in
-        // a digit), not just by a digit. tsc emits both TS6188 at the first
-        // `_` AND TS6189 at the second `_` for these inputs.
-        let ts6188 = diagnostic_codes::NUMERIC_SEPARATORS_ARE_NOT_ALLOWED_HERE;
-        let ts6189 = diagnostic_codes::MULTIPLE_CONSECUTIVE_NUMERIC_SEPARATORS_ARE_NOT_PERMITTED;
-        assert_eq!(
-            scan_separator_diagnostics("0__0.0e0"),
-            vec![(1, ts6188), (2, ts6189)]
-        );
-        assert_eq!(
-            scan_separator_diagnostics("0__0.0e+0"),
-            vec![(1, ts6188), (2, ts6189)]
-        );
-        assert_eq!(
-            scan_separator_diagnostics("0__0.0e-0"),
-            vec![(1, ts6188), (2, ts6189)]
-        );
-    }
-}
+mod tests;
